旧画面と計算ロジックを分離
