+++ docs/design-principles.md
@@ -52,6 +52,13 @@
 
 古い年版は当時の比較資料として残し、現行版の規則を遡って適用しません。
 
+## コード構成
+
+- `App.tsx`は現在表示する画面だけを持ち、旧画面をアーカイブ目的で残しません。
+- 路線データと運賃・料金計算はReactコンポーネントから分離し、画面から計算モジュールを利用します。
+- 2022年版の計算コードは参考表示に必要な現役の計算モジュールとして保持し、使用していないUIとは共存させません。
+- 画面内の「区間指定」と「ランキング」はページルーティングではなくタブとして扱い、ホスティングサービス固有のrewriteを必要としない構成にします。
+
 ## 計算根拠
 
 本番の計算ロジックとデータは、次の一次資料から導出します。
