+++ package.json
@@ -4,12 +4,9 @@
   "private": true,
   "type": "module",
   "dependencies": {
-    "bootstrap": "5.3.8",
-    "bootstrap-icons": "^1.13.1",
     "csv-parse": "^7.0.1",
     "lodash": "4.18.1",
     "react": "19.2.8",
-    "react-bootstrap": "2.10.10",
     "react-dom": "19.2.8"
   },
   "devDependencies": {
