+++ src/App.css
@@ -1,38 +1,107 @@
-.App {
-  text-align: center;
+:root {
+  color: #13231d;
+  background: #f2f4ef;
+  font-family: "Hiragino Kaku Gothic ProN", "Yu Gothic", system-ui, sans-serif;
+  font-synthesis: none;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+* { box-sizing: border-box; }
+body { margin: 0; min-width: 320px; background: #f2f4ef; }
+button, select, input { font: inherit; }
+button, select { color: inherit; }
+
+.app-shell { min-height: 100vh; }
+.site-header { height: 78px; padding: 0 5vw; display: flex; align-items: center; justify-content: space-between; border-bottom: 1px solid #bcc5bd; background: rgba(242, 244, 239, .94); position: sticky; top: 0; z-index: 20; backdrop-filter: blur(12px); }
+.site-header h1 { margin: 0; font-size: 20px; letter-spacing: .04em; }
+.site-header a { color: #13231d; text-decoration: none; border-bottom: 1px solid; font-size: 13px; }
+
+main { width: min(1180px, 92vw); margin: 0 auto; }
+
+.tabs { display: flex; gap: 32px; border-bottom: 1px solid #bcc5bd; padding-top: 18px; }
+.tabs button { appearance: none; background: none; border: 0; padding: 22px 0 18px; color: #65746b; cursor: pointer; position: relative; }
+.tabs button.active { color: #13231d; font-weight: 700; }
+.tabs button.active::after { content: ""; position: absolute; height: 4px; background: #ff5a35; left: 0; right: 0; bottom: -1px; }
+
+.control-bar { display: grid; grid-template-columns: 1.25fr 1fr 1fr; gap: 1px; background: #bcc5bd; border: 1px solid #bcc5bd; margin: 28px 0; }
+.control-bar label, .field-grid label { background: #fff; padding: 13px 16px 10px; color: #65746b; font-size: 11px; font-weight: 700; }
+select { width: 100%; border: 0; border-bottom: 1px solid #94a198; background: transparent; padding: 8px 24px 8px 0; margin-top: 3px; font-size: 14px; font-weight: 700; border-radius: 0; }
+
+.dashboard-grid { display: grid; grid-template-columns: 1fr; gap: 24px; }
+.panel { background: #fff; border: 1px solid #bcc5bd; }
+.journey-panel, .result-panel, .ranking-panel { padding: clamp(20px, 3vw, 38px); }
+.result-panel { position: static; }
+.panel-heading { margin-bottom: 28px; }
+.panel-heading h3, .section-title h3 { margin: 0; font-size: 24px; }
+.field-grid { display: grid; grid-template-columns: repeat(2, 1fr); gap: 12px; margin-bottom: 18px; }
+.field-grid label { border: 1px solid #ccd3cd; padding: 10px 12px; }
+
+.range-select { display: grid; grid-template-columns: 1fr 1.3fr auto 1.3fr; align-items: end; gap: 10px; padding: 15px 0; }
+.range-select > span:first-child { color: #65746b; font-size: 12px; font-weight: 700; align-self: center; }
+.range-arrow { color: #ff5a35; align-self: center; }
+.segment-control { border-top: 1px solid #dce1dc; padding-top: 16px; margin-top: 10px; }
+.switch { display: flex; gap: 10px; align-items: center; font-weight: 700; cursor: pointer; }
+.switch input { width: 18px; height: 18px; accent-color: #087a52; }
+.facility-picker { border: 0; border-top: 1px solid #dce1dc; padding: 18px 0 0; margin: 18px 0 0; display: grid; grid-template-columns: repeat(2, 1fr); gap: 10px; }
+.facility-picker legend { grid-column: 1 / -1; padding: 0 0 12px; font-size: 13px; font-weight: 700; }
+.facility-option { border: 1px solid #ccd3cd; padding: 14px; display: grid; grid-template-columns: auto 1fr; column-gap: 10px; cursor: pointer; }
+.facility-option:has(input:checked) { border-color: #087a52; background: #eef8f2; box-shadow: inset 0 0 0 1px #087a52; }
+.facility-option:has(input:disabled) { cursor: not-allowed; opacity: .55; }
+.facility-option input { width: 18px; height: 18px; margin: 2px 0 0; accent-color: #087a52; grid-row: 1 / 3; }
+.facility-option strong { font-size: 14px; }
+.facility-option span { color: #65746b; font-size: 11px; margin-top: 4px; }
+.constraint-note { color: #65746b; border-top: 1px solid #dce1dc; padding-top: 16px; margin: 20px 0 0; font-size: 12px; line-height: 1.7; }
+
+.result-route { display: flex; justify-content: space-between; align-items: baseline; gap: 16px; margin-bottom: 14px; }
+.result-route strong { font-size: clamp(22px, 4vw, 36px); letter-spacing: -.04em; }
+.result-route span { color: #65746b; font: 700 12px ui-monospace, monospace; white-space: nowrap; }
+.result-section { margin-top: 24px; }
+.result-section h4 { margin: 0 0 10px; font-size: 15px; }
+.points-total { background: #087a52; color: #fff; padding: 22px; display: flex; flex-direction: column; gap: 10px; }
+.points-total span { font-size: 12px; }
+.points-total strong { font: 800 clamp(42px, 5vw, 68px)/1 ui-monospace, monospace; letter-spacing: -.06em; }
+.points-total small { font-size: 16px; letter-spacing: 0; }
+.rate-grid { display: grid; grid-template-columns: 1fr 1fr; gap: 1px; background: #bcc5bd; border: 1px solid #bcc5bd; }
+.rate-grid.single { grid-template-columns: 1fr; }
+.rate-card { background: #f8faf6; padding: 16px; }
+.rate-card span { display: block; color: #65746b; font-size: 11px; }
+.rate-card strong { display: block; margin-top: 9px; font: 800 19px ui-monospace, monospace; }
+.rate-card small { display: block; margin-top: 4px; color: #65746b; font: 700 12px ui-monospace, monospace; }
+.breakdown { margin: 0; }
+.breakdown div { display: flex; justify-content: space-between; padding: 10px 0; border-bottom: 1px solid #dce1dc; }
+.breakdown dt { color: #65746b; }
+.breakdown dd { margin: 0; font-weight: 700; }
+.breakdown .total { border-bottom: 0; padding-top: 18px; font-size: 18px; }
+.breakdown .total dt { color: #13231d; font-weight: 700; }
+.unavailable { padding: 28px; background: #fff2e9; border-left: 4px solid #ff5a35; }
+.unavailable p { color: #6a5b51; line-height: 1.7; }
+
+.ranking-note { color: #65746b; line-height: 1.7; }
+.ranking-tools { max-width: 900px; margin-bottom: 18px; display: grid; grid-template-columns: repeat(3, 1fr); gap: 18px; }
+.ranking-tools label { display: block; color: #65746b; font-size: 11px; font-weight: 700; }
+.ranking-table-wrap { overflow-x: auto; }
+.ranking-table { width: 100%; border-collapse: collapse; min-width: 720px; }
+.ranking-table.ordinary { min-width: 980px; }
+.ranking-table th { text-align: left; color: #65746b; font-size: 11px; letter-spacing: .08em; border-bottom: 2px solid #13231d; padding: 12px 10px; }
+.ranking-table td { border-bottom: 1px solid #dce1dc; padding: 15px 10px; font: 13px ui-monospace, monospace; }
+.ranking-table td:nth-child(2) { font-family: inherit; }
+.ranking-table td small { display: block; color: #65746b; margin-bottom: 3px; }
+.ranking-table td strong { display: block; }
+.ranking-table .selected-rate { color: #087a52; font-size: 18px; }
+
+.notice { border: 1px solid #cf5b36; background: #fff3eb; padding: 20px 24px; margin-bottom: 56px; display: grid; grid-template-columns: 160px 1fr; gap: 20px; }
+.notice p { margin: 0; color: #624f45; line-height: 1.7; }
+@media (max-width: 820px) {
+  .control-bar, .dashboard-grid { grid-template-columns: 1fr; }
+  .result-panel { position: static; }
+  .control-bar { gap: 1px; }
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+@media (max-width: 560px) {
+  .site-header { height: 66px; }
+  .control-bar { margin: 20px 0; }
+  .field-grid, .rate-grid, .facility-picker, .ranking-tools { grid-template-columns: 1fr; }
+  .range-select { grid-template-columns: 1fr auto 1fr; }
+  .range-select > span:first-child { grid-column: 1 / -1; }
+  .journey-panel, .result-panel, .ranking-panel { padding: 20px; }
+  .notice { grid-template-columns: 1fr; }
 }
