+++ src/App.tsx
@@ -1,2720 +1,901 @@
-import "bootstrap-icons/font/bootstrap-icons.css";
-import "bootstrap/dist/css/bootstrap.min.css";
-import { parse } from "csv-parse/browser/esm/sync";
-import { ceil, round, sum } from "lodash";
-import type * as React from "react";
-import type { Reducer } from "react";
+import { useMemo, useState } from "react";
 import {
-  Fragment,
-  useContext,
-  useMemo,
-  useReducer,
-  useState,
-} from "react";
+  legacy2022Engine,
+  type ExpressTicket,
+  type Line,
+  type Season,
+  type SortedSection,
+  type Station,
+  type TotalFare,
+} from "./legacy2022Engine";
+import type {
+  DataVersion,
+  Facility,
+  Interval,
+  JourneySelection,
+} from "./domain/types";
+import { highestFacility } from "./domain/types";
+import { get2022Points } from "./domain/versions/2022";
 import {
-  Accordion,
-  AccordionContext,
-  Alert,
-  Badge,
-  Card,
-  Col,
-  Container,
-  Dropdown,
-  DropdownButton,
-  Fade,
-  FloatingLabel,
-  Form,
-  Nav,
-  Navbar,
-  OverlayTrigger,
-  Popover,
-  Row,
-  Table,
-  ToggleButton,
-} from "react-bootstrap";
+  get2026EastBasicFare,
+  get2026ExpressFare,
+  get2026JourneyPoints,
+  get2026ShinshuPreDcPoints,
+  get2026SpecialVehicleFare,
+} from "./domain/versions/2026";
 import "./App.css";
 
-interface Station {
-  readonly index: number;
-  readonly name: string;
-  /**
-   * 起点からの営業キロ
-   */
-  readonly distance: number;
-}
-
-/**
- * 特定都区市内または東京山手線内
- */
-interface Zone {
-  readonly name: string;
-  readonly central: Station;
-  readonly stations: ReadonlySet<Station>;
-}
-
-type Line = readonly Station[];
-
-interface LineGroup {
-  readonly name: string;
-  readonly lines: readonly Line[];
-}
-
-// (\d+(?:\.\d)?)\t(.+)
-/** 東北新幹線 */
-const line0: Line = [
-  { name: "東京", distance: 0 },
-  { name: "上野", distance: 3.6 },
-  { name: "大宮", distance: 30.3 },
-  { name: "小山", distance: 80.6 },
-  { name: "宇都宮", distance: 109.5 },
-  { name: "那須塩原", distance: 157.8 },
-  { name: "新白河", distance: 185.4 },
-  { name: "郡山", distance: 226.7 },
-  { name: "福島", distance: 272.8 },
-  { name: "白石蔵王", distance: 306.8 },
-  { name: "仙台", distance: 351.8 },
-  { name: "古川", distance: 395 },
-  { name: "くりこま高原", distance: 416.2 },
-  { name: "一ノ関", distance: 445.1 },
-  { name: "水沢江刺", distance: 470.1 },
-  { name: "北上", distance: 487.5 },
-  { name: "新花巻", distance: 500 },
-  { name: "盛岡", distance: 535.3 },
-  { name: "いわて沼宮内", distance: 566.4 },
-  { name: "二戸", distance: 601 },
-  { name: "八戸", distance: 631.9 },
-  { name: "七戸十和田", distance: 668 },
-  { name: "新青森", distance: 713.7 },
-].map((value, index) => ({ ...value, index }));
-
-/** 秋田新幹線 */
-const line1: Line = [
-  ...line0.slice(
-    line0.findIndex(({ name }) => name === "東京"),
-    line0.findIndex(({ name }) => name === "盛岡") + 1
-  ),
-  ...[
-    { name: "雫石", distance: 551.3 },
-    { name: "田沢湖", distance: 575.4 },
-    { name: "角館", distance: 594.1 },
-    { name: "大曲", distance: 610.9 },
-    { name: "秋田", distance: 662.6 },
-  ].map((value, index) => ({
-    ...value,
-    index: index + line0.findIndex(({ name }) => name === "盛岡") + 1,
-  })),
-];
-
-/** 山形新幹線 */
-const line2: Line = [
-  ...line0.slice(
-    line0.findIndex(({ name }) => name === "東京"),
-    line0.findIndex(({ name }) => name === "福島") + 1
-  ),
-  ...[
-    { name: "米沢", distance: 312.9 },
-    { name: "高畠", distance: 322.7 },
-    { name: "赤湯", distance: 328.9 },
-    { name: "かみのやま温泉", distance: 347.8 },
-    { name: "山形", distance: 359.9 },
-    { name: "天童", distance: 373.2 },
-    { name: "さくらんぼ東根", distance: 380.9 },
-    { name: "村山", distance: 386.3 },
-    { name: "大石田", distance: 399.7 },
-    { name: "新庄", distance: 421.4 },
-  ].map((value, index) => ({
-    ...value,
-    index: index + line0.findIndex(({ name }) => name === "福島") + 1,
-  })),
-];
-
-/** 上越新幹線（新潟方面） */
-const line3: Line = [
-  ...line0.slice(
-    line0.findIndex(({ name }) => name === "東京"),
-    line0.findIndex(({ name }) => name === "大宮") + 1
-  ),
-  ...[
-    { name: "熊谷", distance: 64.7 },
-    { name: "本庄早稲田", distance: 86 },
-    { name: "高崎", distance: 105 },
-    { name: "上毛高原", distance: 151.6 },
-    { name: "越後湯沢", distance: 199.2 },
-    { name: "浦佐", distance: 228.9 },
-    { name: "長岡", distance: 270.6 },
-    { name: "燕三条", distance: 293.8 },
-    { name: "新潟", distance: 333.9 },
-  ].map((value, index) => ({
-    ...value,
-    index: index + line0.findIndex(({ name }) => name === "大宮") + 1,
-  })),
-];
-
-/** 上越新幹線（ガーラ湯沢方面） */
-const line4: Line = [
-  ...line3.slice(
-    line3.findIndex(({ name }) => name === "東京"),
-    line3.findIndex(({ name }) => name === "越後湯沢") + 1
-  ),
-  {
-    name: "ガーラ湯沢",
-    distance: 201.0,
-    index: line3.findIndex(({ name }) => name === "越後湯沢") + 1,
-  },
-];
-
-/** 北陸新幹線 */
-const line5 = [
-  ...line0.slice(
-    line0.findIndex(({ name }) => name === "東京"),
-    line0.findIndex(({ name }) => name === "大宮") + 1
-  ),
-  ...line3.slice(
-    line3.findIndex(({ name }) => name === "熊谷"),
-    line3.findIndex(({ name }) => name === "高崎") + 1
-  ),
-  ...[
-    { name: "安中榛名", distance: 123.5 },
-    { name: "軽井沢", distance: 146.8 },
-    { name: "佐久平", distance: 164.4 },
-    { name: "上田", distance: 189.2 },
-    { name: "長野", distance: 222.4 },
-    { name: "飯山", distance: 252.3 },
-    { name: "上越妙高", distance: 281.9 },
-    // { name: "糸魚川", distance: 318.9 },
-    // { name: "黒部宇奈月温泉", distance: 358.1 },
-    // { name: "富山", distance: 391.9 },
-    // { name: "新高岡", distance: 410.8 },
-    // { name: "金沢", distance: 450.5 },
-  ].map((value, index) => ({
-    ...value,
-    index: index + line3.findIndex(({ name }) => name === "高崎") + 1,
-  })),
-];
-
-const lineGroups: ReadonlyMap<string, LineGroup> = new Map(
-  [
-    { name: "東北新幹線", lines: [line0] },
-    { name: "秋田新幹線", lines: [line1] },
-    { name: "山形新幹線", lines: [line2] },
-    { name: "上越新幹線", lines: [line3, line4] },
-    { name: "北陸新幹線", lines: [line5] },
-  ].map((group) => [group.name, group])
-);
-
-const zone0: Zone = {
-  name: "東京山手線内",
-  central: line0.find(({ name }) => name === "東京")!,
-  stations: new Set([
-    line0.find(({ name }) => name === "東京")!,
-    line0.find(({ name }) => name === "上野")!,
-  ]),
-};
-
-const zone1: Zone = {
-  name: "東京都区内",
-  central: line0.find(({ name }) => name === "東京")!,
-  stations: new Set([
-    line0.find(({ name }) => name === "東京")!,
-    line0.find(({ name }) => name === "上野")!,
-  ]),
-};
-
-const zone2: Zone = {
-  name: "仙台市内",
-  central: line0.find(({ name }) => name === "仙台")!,
-  stations: new Set([line0.find(({ name }) => name === "仙台")!]),
-};
-
-const cityZones: readonly Zone[] = [zone1, zone2];
-
-const highSpeedTrains: ReadonlyMap<Line, string> = new Map([
-  [line0, "はやぶさ"],
-  [line1, "こまち"],
-]);
-
-const junctions: ReadonlyMap<Line, Station> = new Map(
-  (
-    [
-      [line1, "盛岡"],
-      [line2, "福島"],
-      [line3, "大宮"],
-      [line4, "越後湯沢"],
-      [line5, "高崎"],
-    ] as const
-  ).map(([line, station]) => [line, line.find(({ name }) => name === station)!])
-);
-
-type FareTable = readonly { readonly [column: string]: string }[];
-
-const csvs: readonly (readonly [Line, string])[] = [
-  [
-    line0,
-    `
-駅名,東京,上野,大宮,小山,宇都宮,那須塩原,新白河,郡山,福島,白石蔵王,仙台,古川,くりこま高原,一ノ関,水沢江刺,北上,新花巻,盛岡,いわて沼宮内,二戸,八戸,七戸十和田
-上野,2400,,,,,,,,,,,,,,,,,,,,,
-大宮,2610,2400,,,,,,,,,,,,,,,,,,,,
-小山,2610,2400,2400,,,,,,,,,,,,,,,,,,,
-宇都宮,3040,2830,2400,2400,,,,,,,,,,,,,,,,,,
-那須塩原,3380,3170,3170,2400,2400,,,,,,,,,,,,,,,,,
-新白河,3380,3170,3170,3170,2400,2400,,,,,,,,,,,,,,,,
-郡山,4270,4060,3170,3170,3170,2400,2400,,,,,,,,,,,,,,,
-福島,4270,4060,4060,3170,3170,3170,2400,2400,,,,,,,,,,,,,,
-白石蔵王,5040,4830,4060,4060,3170,3170,3170,2400,2400,,,,,,,,,,,,,
-仙台,5040,4830,4830,4060,4060,3170,3170,3170,2400,2400,,,,,,,,,,,,
-古川,5040,4830,4830,4830,4060,4060,4060,3170,3170,2400,2400,,,,,,,,,,,
-くりこま高原,5580,5370,4830,4830,4830,4060,4060,3170,3170,3170,2400,2400,,,,,,,,,,
-一ノ関,5580,5370,5370,4830,4830,4060,4060,4060,3170,3170,2400,2400,2400,,,,,,,,,
-水沢江刺,5580,5370,5370,4830,4830,4830,4060,4060,3170,3170,3170,2400,2400,2400,,,,,,,,
-北上,5580,5370,5370,5370,4830,4830,4830,4060,4060,3170,3170,2400,2400,2400,2400,,,,,,,
-新花巻,5580,5370,5370,5370,4830,4830,4830,4060,4060,3170,3170,3170,2400,2400,2400,2400,,,,,,
-盛岡,5910,5700,5370,5370,5370,4830,4830,4830,4060,4060,3170,3170,3170,2400,2400,2400,2400,,,,,
-いわて沼宮内,5910,5700,5700,5370,5370,5370,4830,4830,4060,4060,4060,3170,3170,3170,2400,2400,2400,2400,,,,
-二戸,5910,5700,5700,5700,5370,5370,5370,4830,4830,4060,4060,4060,3170,3170,3170,3170,3170,2400,2400,,,
-八戸,6280,6070,6070,5700,5700,5370,5370,5370,4830,4830,4060,4060,4060,3170,3170,3170,3170,2400,2400,2400,,
-七戸十和田,6280,6070,6070,5700,5700,5700,5370,5370,4830,4830,4830,4060,4060,4060,3170,3170,3170,3170,3170,2400,2400,
-新青森,6810,6600,6070,6070,6070,5700,5700,5370,5370,5370,4830,4830,4060,4060,4060,4060,4060,3170,3170,3170,2400,2400
-`,
-  ],
-  [
-    line3,
-    `
-駅名,東京,上野,大宮,熊谷,本庄早稲田,高崎,上毛高原,越後湯沢,浦佐,長岡,燕三条
-上野,2400,,,,,,,,,,
-大宮,2610,2400,,,,,,,,,
-熊谷,2610,2400,2400,,,,,,,,
-本庄早稲田,2610,2400,2400,2400,,,,,,,
-高崎,3040,2830,2400,2400,2400,,,,,,
-上毛高原,3380,3170,3170,2400,2400,2400,,,,,
-越後湯沢,3380,3170,3170,3170,3170,2400,2400,,,,
-浦佐,4270,4060,3170,3170,3170,3170,2400,2400,,,
-長岡,4270,4060,4060,4060,3170,3170,3170,2400,2400,,
-燕三条,4270,4060,4060,4060,4060,3170,3170,2400,2400,2400,
-新潟,5040,4830,4830,4060,4060,4060,3170,3170,3170,2400,2400
-`,
-  ],
-  [
-    line5,
-    `
-駅名,東京,上野,大宮,熊谷,本庄早稲田,高崎,安中榛名,軽井沢,佐久平,上田,長野,飯山,上越妙高,糸魚川,黒部宇奈月温泉,富山,新高岡
-上野,2400,,,,,,,,,,,,,,,,
-大宮,2610,2400,,,,,,,,,,,,,,,
-熊谷,2610,2400,2400,,,,,,,,,,,,,,
-本庄早稲田,2610,2400,2400,2400,,,,,,,,,,,,,
-高崎,3040,2830,2400,2400,2400,,,,,,,,,,,,
-安中榛名,3040,2830,2400,2400,2400,2400,,,,,,,,,,,
-軽井沢,3380,3170,3170,2400,2400,2400,2400,,,,,,,,,,
-佐久平,3380,3170,3170,2400,2400,2400,2400,2400,,,,,,,,,
-上田,3380,3170,3170,3170,3170,2400,2400,2400,2400,,,,,,,,
-長野,4270,4060,3170,3170,3170,3170,2400,2400,2400,2400,,,,,,,
-飯山,4270,4060,4060,3170,3170,3170,3170,3170,2400,2400,2400,,,,,,
-上越妙高,4270,4060,4060,4060,3170,3170,3170,3170,3170,2400,2400,2400,,,,,
-糸魚川,5700,5490,4730,4730,4730,4730,3830,3830,3830,3830,3070,3070,2400,,,,
-黒部宇奈月温泉,6030,5820,5820,5050,5050,5050,5050,5050,4160,4160,4160,3830,2400,2400,,,
-富山,6360,6150,6150,6150,6150,5390,5390,5390,5390,5390,4160,3830,3170,2400,2400,,
-新高岡,6900,6690,6150,6150,6150,6150,5390,5390,5390,5390,4160,3830,3170,2400,2400,2400,
-金沢,6900,6690,6690,6150,6150,6150,6150,6150,5390,5390,5050,3830,3170,3170,2400,2400,2400
-`,
-  ],
-];
-
-const tables: ReadonlyMap<Line, FareTable> = new Map(
-  csvs.map(
-    ([line, raw]) => [line, parse(raw.trim(), { columns: true })] as const
-  )
-);
-
-const table: FareTable = parse(
-  `
-駅名,東京,上野,大宮,仙台,古川,くりこま高原,一ノ関,水沢江刺,北上,新花巻,盛岡,いわて沼宮内,二戸,八戸,七戸十和田
-上野,2400,,,,,,,,,,,,,,
-大宮,2610,2400,,,,,,,,,,,,,
-仙台,5360,5150,5150,,,,,,,,,,,,
-古川,5360,5150,5150,2500,,,,,,,,,,,
-くりこま高原,6000,5790,5250,2500,2500,,,,,,,,,,
-一ノ関,6000,5790,5790,2500,2500,2500,,,,,,,,,
-水沢江刺,6000,5790,5790,3380,2500,2500,2500,,,,,,,,
-北上,6000,5790,5790,3380,2500,2500,2500,2500,,,,,,,
-新花巻,6000,5790,5790,3380,3380,2500,2500,2500,2500,,,,,,
-盛岡,6430,6220,5890,3380,3380,3380,2500,2500,2500,2500,,,,,
-いわて沼宮内,6430,6220,6220,4270,3380,3380,3270,2500,2500,2500,2400,,,,
-二戸,6430,6220,6220,4270,4270,3380,3270,3270,3270,3270,2400,2400,,,
-八戸,6800,6590,6590,4270,4270,4270,3270,3270,3270,3270,2400,2400,2400,,
-七戸十和田,6800,6590,6590,5040,4270,4270,4160,3270,3270,3270,3170,3170,2400,2400,
-新青森,7330,7120,6590,5040,5040,4270,4160,4160,4160,4160,3170,3170,3170,2400,2400
-`.trim(),
-  { columns: true }
-);
-
-const StationDropdown: React.FC<{
-  value?: Station;
-  header: string;
-  onChange: (value: Station) => void;
-  items: readonly { station: Station; disabled: boolean; active: boolean }[];
-}> = ({ value, header, onChange, items }) => {
-  return (
-    <DropdownButton
-      title={value?.name ?? header}
-      variant="secondary"
-      className="d-grid"
-    >
-      <Dropdown.Header>{header}</Dropdown.Header>
-      {items.map(({ station, disabled, active }) => (
-        <Dropdown.Item
-          onClick={() => onChange(station)}
-          className="d-flex justify-content-between"
-          key={station.name}
-          disabled={disabled}
-          active={active}
-        >
-          <span>{station.name}</span>
-          <code className="ms-5">{station.distance.toFixed(1)}</code>
-        </Dropdown.Item>
-      ))}
-    </DropdownButton>
-  );
-};
-
-/**
- * 地方交通線の営業キロの区間
- * @param distance 10キロメートルを超え、1200キロメートルまでの営業キロ
- * @returns 中央の営業キロ
- */
-const getDistance2 = (distance: number) =>
-  distance <= 15
-    ? 13
-    : distance <= 20
-    ? 18
-    : distance <= 23
-    ? 22
-    : distance <= 28
-    ? 26
-    : distance <= 32
-    ? 30
-    : distance <= 37
-    ? 35
-    : distance <= 41
-    ? 39
-    : distance <= 46
-    ? 44
-    : distance <= 55
-    ? 51
-    : distance <= 64
-    ? 60
-    : distance <= 73
-    ? 69
-    : distance <= 82
-    ? 78
-    : distance <= 91
-    ? 87
-    : distance <= 100
-    ? 96
-    : distance <= 110
-    ? 105
-    : distance <= 128
-    ? 119
-    : distance <= 146
-    ? 137
-    : distance <= 164
-    ? 155
-    : distance <= 182
-    ? 173
-    : distance <= 200
-    ? 191
-    : distance <= 219
-    ? 210
-    : distance <= 237
-    ? 228
-    : distance <= 255
-    ? 246
-    : distance <= 273
-    ? 264
-    : distance <= 291
-    ? 282
-    : distance <= 310
-    ? 301
-    : distance <= 328
-    ? 319
-    : distance <= 346
-    ? 337
-    : distance <= 364
-    ? 355
-    : distance <= 382
-    ? 373
-    : distance <= 400
-    ? 391
-    : distance <= 419
-    ? 410
-    : distance <= 437
-    ? 428
-    : distance <= 455
-    ? 446
-    : distance <= 473
-    ? 464
-    : distance <= 491
-    ? 482
-    : distance <= 510
-    ? 501
-    : distance <= 528
-    ? 519
-    : distance <= 546
-    ? 537
-    : distance <= 582
-    ? 564
-    : distance <= 619
-    ? 601
-    : distance <= 655
-    ? 637
-    : distance <= 691
-    ? 673
-    : distance <= 728
-    ? 710
-    : distance <= 764
-    ? 746
-    : distance <= 800
-    ? 782
-    : distance <= 837
-    ? 819
-    : distance <= 873
-    ? 855
-    : distance <= 910
-    ? 892
-    : distance <= 946
-    ? 928
-    : distance <= 982
-    ? 964
-    : distance <= 1019
-    ? 1001
-    : distance <= 1055
-    ? 1037
-    : distance <= 1091
-    ? 1073
-    : distance <= 1128
-    ? 1110
-    : distance <= 1164
-    ? 1146
-    : 1182;
-
-/**
- *
- * @param distance 10キロメートルを超える営業キロ
- * @returns
- */
-const getDistance1 = (distance: number) =>
-  distance > 600
-    ? Math.ceil(distance / 40) * 40 - 20
-    : distance > 100
-    ? Math.ceil(distance / 20) * 20 - 10
-    : distance > 50
-    ? Math.ceil(distance / 10) * 10 - 5
-    : Math.ceil(distance / 5) * 5 - 2;
-
-/**
- * 指定した区間の距離を返す。丸め誤差は取り除く。
- * @param section 距離を求める区間
- * @returns 距離
- */
-const getDistance0 = (section: SortedSection) =>
-  round(section.arrival.distance - section.departure.distance, 1);
-
-/**
- * **幹線**内相互発着となる場合の大人片道普通旅客運賃を計算する
- * @param distance 営業キロ
- * @returns 運賃
- */
-const getBasicFare0 = (distance: number) => {
-  // 1-3: 150, 4-6: 190, 7-10: 200
-  const distance1 = getDistance1(distance);
-  const fare0 =
-    16.2 * Math.min(distance1, 300) +
-    12.85 * Math.max(Math.min(distance1 - 300, 300), 0) +
-    7.05 * Math.max(distance1 - 600, 0);
-
-  return distance > 10
-    ? round((distance > 100 ? round(fare0, -2) : ceil(fare0, -1)) * 1.1, -1)
-    : distance > 6
-    ? 200
-    : distance > 3
-    ? 190
-    : 150;
-};
-
-/**
- * **地方交通線**内相互発着となる場合の大人片道普通旅客運賃を計算する
- * @param distance 営業キロ
- * @returns 運賃
- */
-const getBasicFare1 = (distance: number) => {
-  // 1-3: 150, 4-6: 190, 7-10: 210
-  const distance2 = getDistance2(distance);
-  const fare0 =
-    17.8 * Math.min(distance2, 273) +
-    14.1 * Math.max(Math.min(distance2 - 273, 273), 0) +
-    7.7 * Math.max(distance2 - 546, 0);
-
-  return 10 < distance && distance <= 15
-    ? 240
-    : 15 < distance && distance <= 20
-    ? 330
-    : 20 < distance && distance <= 23
-    ? 420
-    : 23 < distance && distance <= 28
-    ? 510
-    : 32 < distance && distance <= 37
-    ? 680
-    : 41 < distance && distance <= 46
-    ? 860
-    : 46 < distance && distance <= 55
-    ? 990
-    : 55 < distance && distance <= 64
-    ? 1170
-    : 64 < distance && distance <= 73
-    ? 1340
-    : 73 < distance && distance <= 82
-    ? 1520
-    : 82 < distance && distance <= 91
-    ? 1690
-    : 100 < distance && distance <= 110
-    ? 1980
-    : 291 < distance && distance <= 310
-    ? 5720
-    : distance > 10
-    ? round((distance > 100 ? round(fare0, -2) : ceil(fare0, -1)) * 1.1, -1)
-    : distance > 6
-    ? 210
-    : distance > 3
-    ? 190
-    : 150;
-};
-
-/**
- * **東京附近における電車特定区間**内相互発着の場合の大人片道普通旅客運賃を計算する
- * @param distance 営業キロ
- * @returns 運賃
- */
-const getBasicFare2 = (distance: number) => {
-  // 1-3: 140, 4-6: 160, 7-10: 170
-  const distance1 = getDistance1(distance);
-  const fare0 =
-    15.3 * Math.min(distance1, 300) +
-    12.15 * Math.max(Math.min(distance1 - 300, 300), 0);
-
-  return distance > 10
-    ? ceil((distance > 100 ? round(fare0, -2) : ceil(fare0, -1)) * 1.1, -1)
-    : distance > 6
-    ? 170
-    : distance > 3
-    ? 160
-    : 140;
-};
-
-const reserved = "指定席";
-const nonReserved = "自由席";
-const specific = "特定";
-const standingOnly = "立席";
-
-/**
- * 特急券
- */
-interface ExpressTicket {
-  /**
-   * 全乗車区間
-   */
-  readonly section: Section;
-  /**
-   * 特急料金
-   */
-  readonly fare: number;
-  /**
-   * はやぶさ号やこまち号を利用する区間
-   */
-  readonly highSpeed?: Section | undefined;
-  /**
-   * 特急券の種類
-   */
-  readonly type: typeof reserved | typeof nonReserved | typeof specific;
-  /**
-   * 利用可能な座席の種類
-   */
-  readonly availableSeat:
-    | typeof reserved
-    | typeof nonReserved
-    | typeof standingOnly;
-}
-
-/**
- * 奥羽本線中福島・新庄間並びに田沢湖線及び奥羽本線中大曲・秋田間を、東北新幹線にまたがって利用する場合の指定席特急料金を計算する
- * @param distance 営業キロ
- * @returns 指定席特急料金
- */
-const getLimitedExpressFare1 = (distance: number) =>
-  getLimitedExpressFare3(distance) - 530;
-
-/**
- * 奥羽本線中福島・新庄間並びに田沢湖線及び奥羽本線中大曲・秋田間の指定席特急料金を計算する
- * @param distance 営業キロ
- * @returns 指定席特急料金
- */
-const getLimitedExpressFare3 = (distance: number) =>
-  distance > 100 ? 2110 : distance > 50 ? 1660 : 1290;
-
-/**
- * 上越線に運転する特別急行列車の越後湯沢・ガーラ湯沢相互間に発売する指定席特急券及び自由席特急券に対する特急料金
- */
-const limitedExpressFares2 = {
-  nonReservedOrStandingOnly: 100,
-  reserved: 100,
-};
-
-/**
- * 奥羽本線中福島・新庄間並びに田沢湖線及び奥羽本線中大曲・秋田間の特急料金を計算する
- * @param distance 営業キロ
- * @returns 指定席特急料金、立席特急料金及び自由席特急料金
- */
-const getLimitedExpressFares0 = (distance: number, season: Season) => {
-  const reserved = getLimitedExpressFare3(distance);
-  return {
-    reserved:
-      season === busiest
-        ? reserved + 400
-        : season === busy
-        ? reserved + 200
-        : season === off
-        ? reserved - 200
-        : reserved,
-    nonReservedOrStandingOnly: reserved - 530,
-  } as const;
-};
-
-/**
- * 奥羽本線中福島・新庄間並びに田沢湖線及び奥羽本線中大曲・秋田間を、東北新幹線にまたがって利用する場合の特急料金を計算する
- * @param distance 営業キロ
- * @returns 指定席特急料金・立席特急料金及び自由席特急料金
- */
-const getLimitedExpressFares1 = (distance: number) =>
-  ({
-    reserved: getLimitedExpressFare1(distance),
-    nonReservedOrStandingOnly: getLimitedExpressFare3(distance) - 530,
-  } as const);
-
-/**
- * 新幹線の指定席特急料金を計算する
- * @param table 指定席特急料金の表
- * @param section 区間。 `arrival` は `departure` より終点に近い駅である必要がある
- * @returns 指定席特急料金
- */
-const getSuperExpressFare = (
-  table: FareTable,
-  section: SortedSection
-): number =>
-  +table.find((row) => row["駅名"] === section.arrival.name)![
-    section.departure.name
-  ]!;
-
-/**
- * 新幹線の指定した区間の特急料金を計算する
- * @param line 東北新幹線、上越新幹線、北陸新幹線のいずれか
- * @param section 特急料金を計算する区間。 `arrival` は `departure` より終点に近い駅である必要がある。
- * @param highSpeed はやぶさ号やこまち号を利用する区間
- * @param season シーズン
- * @returns 自由席特急料金、特定特急料金、指定席特急料金
- */
-const getSuperExpressTickets = (
-  line: Line,
-  section: SortedSection,
-  highSpeed: SortedSection | undefined,
-  season: Season
-): Readonly<{
-  reserved: ExpressTicket;
-  nonReserved: ExpressTicket | undefined;
-  reservedHighSpeed: ExpressTicket | undefined;
-  standingOnly: ExpressTicket | undefined;
-}> => {
-  const { departure, arrival } = section;
-
-  const reservedExpressFare = getSuperExpressFare(tables.get(line)!, section);
-
-  const specificExpressFares =
-    departure.name === "郡山" && arrival.name === "福島"
-      ? {
-          nonReservedOrStandingOnly: 880,
-          reserved:
-            season === busiest
-              ? 1810
-              : season === busy
-              ? 1610
-              : season === off
-              ? 1210
-              : 1410,
-        }
-      : departure.name === "東京" && arrival.name === "大宮"
-      ? { nonReservedOrStandingOnly: 1090 }
-      : arrival.index - departure.index === 1 ||
-        [
-          ["古川", "一ノ関"],
-          ["一ノ関", "北上"],
-          ["北上", "盛岡"],
-          ["熊谷", "高崎"],
-        ].some(([a, b]) =>
-          isEquivalent(section, {
-            departure: line.find(({ name }) => a === name)!,
-            arrival: line.find(({ name }) => b === name)!,
-          })
-        )
-      ? { nonReservedOrStandingOnly: getDistance0(section) > 50 ? 1000 : 880 }
-      : {};
-
-  const reservedExpressTicket: ExpressTicket = {
-    availableSeat: reserved,
-    section,
-    ...("reserved" in specificExpressFares
-      ? {
-          type: specific,
-          fare: specificExpressFares.reserved,
-        }
-      : {
-          type: reserved,
-          fare:
-            season === busiest
-              ? reservedExpressFare + 400
-              : season === busy
-              ? reservedExpressFare + 200
-              : season === off
-              ? reservedExpressFare - 200
-              : reservedExpressFare,
-        }),
-  };
-
-  /**
-   * 自由席が利用可能な区間であれば、 `true`
-   */
-  const nonReservedAvailable =
-    line !== line0 ||
-    arrival.index <= line.findIndex(({ name }) => name === "盛岡");
-
-  /**
-   * 立席が利用可能な区間であれば、 `true`
-   */
-  const standingOnlyAvailable =
-    line === line0 &&
-    departure.index >= line.findIndex(({ name }) => name === "盛岡");
-
-  const nonReservedExpressTicket: ExpressTicket | undefined =
-    nonReservedAvailable
-      ? {
-          availableSeat: nonReserved,
-          section,
-          ...("nonReservedOrStandingOnly" in specificExpressFares
-            ? {
-                type: specific,
-                fare: specificExpressFares.nonReservedOrStandingOnly,
-              }
-            : { type: nonReserved, fare: reservedExpressFare - 530 }),
-        }
-      : undefined;
-
-  const standingOnlyExpressTicket: ExpressTicket | undefined =
-    standingOnlyAvailable
-      ? {
-          type: specific,
-          availableSeat: standingOnly,
-          section,
-          fare:
-            "nonReservedOrStandingOnly" in specificExpressFares
-              ? specificExpressFares.nonReservedOrStandingOnly
-              : reservedExpressFare - 530,
-        }
-      : undefined;
-
-  const reservedHighSpeedFare =
-    highSpeed &&
-    (highSpeed.departure === departure && highSpeed.arrival === arrival
-      ? getSuperExpressFare(table, section)
-      : reservedExpressFare +
-        getSuperExpressFare(table, highSpeed) -
-        getSuperExpressFare(tables.get(line)!, highSpeed));
-
-  const reservedHighSpeedTicket: ExpressTicket | undefined =
-    reservedHighSpeedFare !== undefined
-      ? {
-          type: reserved,
-          availableSeat: reserved,
-          section,
-          highSpeed,
-          fare:
-            season === busiest
-              ? reservedHighSpeedFare + 400
-              : season === busy
-              ? reservedHighSpeedFare + 200
-              : season === off
-              ? reservedHighSpeedFare - 200
-              : reservedHighSpeedFare,
-        }
-      : undefined;
-
-  return {
-    reserved: reservedExpressTicket,
-    nonReserved: nonReservedExpressTicket,
-    standingOnly: standingOnlyExpressTicket,
-    reservedHighSpeed: reservedHighSpeedTicket,
-  };
-};
-
-/**
- * 新幹線以外の線区の指定した区間の特急料金を計算する
- * @param line 秋田新幹線、山形新幹線、上越新幹線（ガーラ湯沢方面）のいずれか
- * @param section 特急料金を計算する区間。 `arrival` は `departure` より終点に近い駅である必要がある。
- */
-const getLimitedExpressFares2 = (
-  line: Line,
-  section: SortedSection,
-  season: Season
-) =>
-  line === line4
-    ? limitedExpressFares2
-    : getLimitedExpressFares0(getDistance0(section), season);
-
-/**
- * 新幹線以外の線区の指定した区間の特急料金を計算する
- * @param line 秋田新幹線、山形新幹線、上越新幹線（ガーラ湯沢方面）のいずれか
- * @param section 特急料金を計算する区間。 `arrival` は `departure` より終点に近い駅である必要がある。
- */
-const getLimitedExpressTickets = (
-  line: Line,
-  section: SortedSection,
-  getLimitedExpressFares: (
-    line: Line,
-    section: SortedSection,
-    season: Season
-  ) => {
-    readonly reserved: number;
-    readonly nonReservedOrStandingOnly: number;
-  },
-  season: Season
-): Readonly<{
-  reserved: ExpressTicket;
-  nonReserved: ExpressTicket | undefined;
-  standingOnly: ExpressTicket | undefined;
-}> => {
-  const {
-    reserved: reservedExpressFare,
-    nonReservedOrStandingOnly: nonReservedOrStandingOnlyExpressFare,
-  } = getLimitedExpressFares(line, section, season);
-
-  const reservedExpressTicket: ExpressTicket = {
-    type: reserved,
-    section,
-    fare: reservedExpressFare,
-    availableSeat: reserved,
-  };
-
-  /**
-   * 自由席が利用可能な区間であれば、 `true`
-   */
-  const nonReservedAvailable = line === line4;
-
-  /**
-   * 立席が利用可能な区間であれば、 `true`
-   */
-  const standingOnlyAvailable = line === line1 || line === line2;
-
-  const nonReservedExpressTicket: ExpressTicket | undefined =
-    nonReservedAvailable
-      ? {
-          availableSeat: nonReserved,
-          type: nonReserved,
-          section,
-          fare: nonReservedOrStandingOnlyExpressFare,
-        }
-      : undefined;
-
-  const standingOnlyExpressTicket: ExpressTicket | undefined =
-    standingOnlyAvailable
-      ? {
-          type: specific,
-          availableSeat: standingOnly,
-          section,
-          fare: nonReservedOrStandingOnlyExpressFare,
-        }
-      : undefined;
-
-  return {
-    reserved: reservedExpressTicket,
-    nonReserved: nonReservedExpressTicket,
-    standingOnly: standingOnlyExpressTicket,
-  };
-};
-
-const ExpressFaresLabel: React.FC<{
-  tickets: readonly ExpressTicket[];
-}> = ({ tickets }) => {
-  const text = jpyNameFormatter.format(sum(tickets.map(({ fare }) => fare)));
-  const badges = [
-    ...new Set(
-      tickets
-        .filter(({ type, availableSeat }) => type !== availableSeat)
-        .map(({ type }) => type)
-    ),
-  ].map((type) => <Badge key={type}>{type}</Badge>);
-
-  return tickets.length > 1 ? (
-    <>
-      <OverlayTrigger
-        overlay={
-          <Popover>
-            <Popover.Header>特急料金の内訳</Popover.Header>
-            <Popover.Body>
-              {tickets.map(({ section, fare, type, availableSeat }) => (
-                <Row key={`${section.departure.name}-${section.arrival.name}`}>
-                  <Col xs="auto">
-                    {section.departure.name} <i className="bi bi-arrow-right" />{" "}
-                    {section.arrival.name}{" "}
-                    {type !== availableSeat ? <Badge>{type}</Badge> : undefined}
-                  </Col>
-                  <Col xs="auto" className="ms-auto">
-                    {jpyNameFormatter.format(fare)}
-                  </Col>
-                </Row>
-              ))}
-            </Popover.Body>
-          </Popover>
-        }
-      >
-        <u style={{ textDecoration: "underline dotted var(--bs-info)" }}>
-          {text}
-        </u>
-      </OverlayTrigger>{" "}
-      {badges}
-    </>
-  ) : (
-    <>
-      {text} {badges}
-    </>
-  );
-};
-
-const SeatsLabel: React.FC<{
-  tickets: readonly ExpressTicket[];
-}> = ({ tickets }) => {
-  const seats = new Set(tickets.map(({ availableSeat }) => availableSeat));
-  const text = [...seats].join("・");
-  return seats.size > 1 ? (
-    <OverlayTrigger
-      overlay={
-        <Popover>
-          <Popover.Header>区間ごとの利用座席</Popover.Header>
-          <Popover.Body>
-            {tickets.map(({ section, availableSeat }) => (
-              <Row key={`${section.departure.name}-${section.arrival.name}`}>
-                <Col xs="auto">
-                  {section.departure.name} <i className="bi bi-arrow-right" />{" "}
-                  {section.arrival.name}
-                </Col>
-                <Col xs="auto" className="ms-auto">
-                  {availableSeat}
-                </Col>
-              </Row>
-            ))}
-          </Popover.Body>
-        </Popover>
-      }
-    >
-      <u style={{ textDecoration: "underline dotted var(--bs-info)" }}>
-        {text}
-      </u>
-    </OverlayTrigger>
-  ) : (
-    <>{text}</>
-  );
-};
-
-/**
- * 指定した区間の運賃を計算する
- * @param line
- * @param section 運賃を計算する区間
- * @returns 運賃
- */
-const getBasicFare = (line: Line, section: SortedSection) => {
-  const { departure, arrival } = section;
-  const distance = getDistance0(section);
-
-  return arrival.index <= line.findIndex(({ name }) => name === "大宮")
-    ? getBasicFare2(distance)
-    : line === line1 &&
-      departure.index >= line.findIndex(({ name }) => name === "盛岡") &&
-      arrival.index <= line.findIndex(({ name }) => name === "大曲")
-    ? getBasicFare1(distance)
-    : getBasicFare0(
-        line === line1 &&
-          ["盛岡", "大曲"]
-            .map((name) => line.findIndex((station) => station.name === name))
-            .some((i) => departure.index < i && i < arrival.index)
-          ? distance +
-              round(
-                getDistance0({
-                  sorted: true,
-                  departure:
-                    departure.index <
-                    line.findIndex(({ name }) => name === "盛岡")
-                      ? line.find(({ name }) => name === "盛岡")!
-                      : departure,
-                  arrival:
-                    line.findIndex(({ name }) => name === "大曲") <
-                    arrival.index
-                      ? line.find(({ name }) => name === "大曲")!
-                      : arrival,
-                }) *
-                  (17.8 / 16.2 - 1),
-                1
-              )
-          : distance
-      );
-};
-
-interface TicketType {
-  readonly name: string;
-  readonly url?: URL;
-  isAvailable(line: Line, expressTickets: readonly ExpressTicket[]): boolean;
-}
-
-const ticketTypes: readonly [TicketType, TicketType, TicketType] = [
-  {
-    name: "紙のきっぷ",
-    isAvailable() {
-      return true;
-    },
-  },
-  {
-    name: "タッチでGo!新幹線",
-    url: new URL("https://www.jreast.co.jp/touchdego/"),
-    isAvailable(line: Line, expressTickets: readonly ExpressTicket[]) {
-      return !(
-        expressTickets.some(
-          ({ availableSeat }) => availableSeat === reserved
-        ) ||
-        ((line === line0 || line === line1) &&
-          expressTickets[0]!.section.departure.index <
-            line.findIndex(({ name }) => name === "盛岡") &&
-          expressTickets.slice(-1)[0]!.section.arrival.index >
-            line.findIndex(({ name }) => name === "盛岡")) ||
-        (line === line2 &&
-          expressTickets[0]!.section.departure.index <
-            line.findIndex(({ name }) => name === "福島") &&
-          expressTickets.slice(-1)[0]!.section.arrival.index >
-            line.findIndex(({ name }) => name === "福島"))
-      );
-    },
-  },
-  {
-    name: "新幹線eチケット",
-    url: new URL("https://www.eki-net.com/top/e-ticket/"),
-    isAvailable(line: Line, expressTickets: readonly ExpressTicket[]) {
-      return (
-        !expressTickets.some(
-          ({ availableSeat }) => availableSeat === standingOnly
-        ) &&
-        (expressTickets[0]!.section.departure !==
-          line.find(({ name }) => name === "東京") ||
-          expressTickets.slice(-1)[0]!.section.arrival !==
-            line.find(({ name }) => name === "上野")) &&
-        (expressTickets[0]!.section.departure !==
-          line.find(({ name }) => name === "越後湯沢") ||
-          expressTickets.slice(-1)[0]!.section.arrival !==
-            line.find(({ name }) => name === "ガーラ湯沢"))
-      );
-    },
-  },
-];
-
-interface TotalFare {
-  /**
-   * 運賃
-   */
-  readonly basicFare: number;
-  /**
-   * 特急券
-   */
-  readonly expressTickets: readonly ExpressTicket[];
-  /**
-   * 割引
-   */
-  readonly discount?: number;
-  /**
-   * 運賃・特急料金・割引の合計
-   */
-  readonly total: number;
-  readonly types: readonly TicketType[];
-}
-
-const chooseOneOrBothTicketType = <
-  F extends {
-    readonly total: number;
-    readonly basicFare: number;
-    readonly types: readonly TicketType[];
-  }
->(
-  a: F,
-  b: F
-): F =>
-  a.total < b.total
-    ? a
-    : a.total > b.total
-    ? b
-    : {
-        ...a,
-        types: [...a.types, ...b.types],
-      };
-
-const isPointAvailable = (line: Line, section: SortedSection) =>
-  !isEquivalent(section, {
-    departure: line.find(({ name }) => name === "東京")!,
-    arrival: line.find(({ name }) => name === "上野")!,
-  }) &&
-  !isEquivalent(section, {
-    departure: line.find(({ name }) => name === "越後湯沢")!,
-    arrival: line.find(({ name }) => name === "ガーラ湯沢")!,
-  }) &&
-  (line !== line5 ||
-    section.arrival.index <= line.findIndex(({ name }) => name === "上越妙高"));
-
-const totalFares = <
-  F extends {
-    readonly basicFare: number;
-    readonly expressTickets: readonly ExpressTicket[];
-    readonly discount?: number;
-  }
->(
-  fares: F
-): F & { readonly total: number } => ({
-  ...fares,
-  total:
-    fares.basicFare +
-    sum(fares.expressTickets.map(({ fare }) => fare)) +
-    (fares.discount ?? 0),
-});
-
-const getFareTotalWithSomeTicketType = (
-  line: Line,
-  basicFares: readonly [number, number],
-  expressTickets: readonly ExpressTicket[]
-) =>
-  [
-    {
-      basicFare: basicFares[1],
-      expressTickets,
-      types: [ticketTypes[0]],
-    },
-    ...(ticketTypes[1].isAvailable(line, expressTickets)
-      ? [
-          {
-            basicFare: basicFares[1],
-            expressTickets,
-            types: [ticketTypes[1]],
-          },
-        ]
-      : []),
-    ...(ticketTypes[2].isAvailable(line, expressTickets)
-      ? [
-          {
-            basicFare: basicFares[0],
-            expressTickets,
-            types: [ticketTypes[2]],
-            ...(expressTickets.some(
-              ({ availableSeat }) => availableSeat === reserved
-            )
-              ? { discount: -200 }
-              : {}),
-          },
-        ]
-      : []),
-  ]
-    .map(totalFares)
-    .reduce(chooseOneOrBothTicketType);
-
-interface PointTicketType {
-  readonly name: string;
-  readonly url?: URL;
-  getPoints(distance: number): number;
-}
-
-const pointTicketTypes: readonly [PointTicketType, PointTicketType] = [
-  {
-    name: "通常",
-    getPoints(distance: number) {
-      return distance > 400
-        ? 12110
-        : distance > 200
-        ? 7940
-        : distance > 100
-        ? 4620
-        : 2160;
-    },
-  },
-  {
-    name: "新幹線YEARスペシャル",
-    url: new URL("https://www.jreast.co.jp/shinkansenyear2022/tokuten_ticket/"),
-    getPoints(distance: number) {
-      return distance > 400
-        ? 6000
-        : distance > 200
-        ? 3900
-        : distance > 100
-        ? 2300
-        : 1000;
-    },
-  },
-];
-
-/**
- * 指定した区間の運賃・特急料金を計算する
- * @param line
- * @param section 運賃・特急料金を計算する区間
- * @param highSpeed はやぶさ号やこまち号を利用する区間
- */
-const getFares = ({
-  line,
-  section,
-  highSpeed,
-  season,
-  getPoints,
-}: {
-  line: Line;
-  section: SortedSection;
-  highSpeed: SortedSection | undefined;
-  season: Season;
-  getPoints: (distance: number) => number;
-}) => {
-  const { departure, arrival } = section;
-
-  const distance = getDistance0(section);
-  const points = isPointAvailable(line, section)
-    ? getPoints(distance)
-    : undefined;
-
-  const junction = junctions.get(line);
-
-  const superExpressTickets =
-    (line === line1 || line === line2 || line === line4) && junction
-      ? departure.index < junction.index
-        ? getSuperExpressTickets(
-            line === line4 ? line3 : line0,
-            junction.index < arrival.index
-              ? { sorted: true, departure, arrival: junction }
-              : section,
-            highSpeed &&
-              (junction.index < highSpeed.arrival.index
-                ? {
-                    sorted: true,
-                    departure: highSpeed.departure,
-                    arrival: junction,
-                  }
-                : highSpeed),
-            season
-          )
-        : undefined
-      : getSuperExpressTickets(line, section, highSpeed, season);
-
-  const limitedExpressFares =
-    (line === line1 || line === line2 || line === line4) &&
-    junction &&
-    arrival.index > junction.index
-      ? departure.index >= junction.index
-        ? getLimitedExpressTickets(
-            line,
-            section,
-            getLimitedExpressFares2,
-            season
-          )
-        : getLimitedExpressTickets(
-            line,
-            { departure: junction, arrival, sorted: true },
-            line === line4
-              ? getLimitedExpressFares2
-              : (_, section) => getLimitedExpressFares1(getDistance0(section)),
-            season
-          )
-      : undefined;
-
-  const nonReservedOrStandingOnlySuperExpressTicket =
-    superExpressTickets?.nonReserved ?? superExpressTickets?.standingOnly;
-  const nonReservedOrStandingOnlyLimitedExpressTicket =
-    limitedExpressFares?.nonReserved ?? limitedExpressFares?.standingOnly;
-
-  const nonReservedOrStandingOnlyAvailable: boolean =
-    (!superExpressTickets || !!nonReservedOrStandingOnlySuperExpressTicket) &&
-    (!limitedExpressFares || !!nonReservedOrStandingOnlyLimitedExpressTicket);
-
-  const nonReservedOrStandingOnlyExpressTickets =
-    nonReservedOrStandingOnlyAvailable
-      ? [
-          nonReservedOrStandingOnlySuperExpressTicket,
-          nonReservedOrStandingOnlyLimitedExpressTicket,
-        ].filter((ticket): ticket is ExpressTicket => ticket !== undefined)
-      : undefined;
-
-  const reservedExpressTickets = [
-    superExpressTickets?.reserved,
-    limitedExpressFares?.reserved,
-  ].filter((ticket): ticket is ExpressTicket => ticket !== undefined);
-
-  const reservedHighSpeedExpressTickets =
-    superExpressTickets?.reservedHighSpeed &&
-    [
-      superExpressTickets?.reservedHighSpeed,
-      limitedExpressFares?.reserved,
-    ].filter((ticket): ticket is ExpressTicket => ticket !== undefined);
-
-  const section200: SortedSection = {
-    sorted: true,
-    departure:
-      cityZones.find(({ stations }) => stations.has(section.departure))
-        ?.central ?? section.departure,
-    arrival:
-      cityZones.find(({ stations }) => stations.has(section.arrival))
-        ?.central ?? section.arrival,
-  };
-  const section100: SortedSection = {
-    sorted: true,
-    departure: zone0.stations.has(section.departure)
-      ? zone0.central
-      : section.departure,
-    arrival: zone0.stations.has(section.arrival)
-      ? zone0.central
-      : section.arrival,
-  };
-
-  const basicFare0 = getBasicFare(line, section);
-
-  const basicFare1 =
-    getDistance0(section200) > 200
-      ? getBasicFare(line, section200)
-      : getDistance0(section100) > 100
-      ? getBasicFare(line, section100)
-      : basicFare0;
-
-  const nonReservedOrStandingOnly: TotalFare | undefined =
-    nonReservedOrStandingOnlyExpressTickets &&
-    getFareTotalWithSomeTicketType(
-      line,
-      [basicFare0, basicFare1],
-      nonReservedOrStandingOnlyExpressTickets
-    );
-  const reserved: TotalFare = getFareTotalWithSomeTicketType(
-    line,
-    [basicFare0, basicFare1],
-    reservedExpressTickets
-  );
-  const reservedHighSpeed: TotalFare | undefined =
-    reservedHighSpeedExpressTickets &&
-    getFareTotalWithSomeTicketType(
-      line,
-      [basicFare0, basicFare1],
-      reservedHighSpeedExpressTickets
-    );
-
-  return {
-    distance,
-    nonReservedOrStandingOnly: nonReservedOrStandingOnly,
-    reserved,
-    reservedHighSpeed,
-    points,
-  } as const;
-};
-
-const isEquivalent = (a: Section, b: Section) =>
-  a.departure === b.departure && a.arrival === b.arrival;
-
-const reverseSection = (section: Section): Section => ({
-  departure: section.arrival,
-  arrival: section.departure,
-});
-const reverseTicket = (ticket: ExpressTicket): ExpressTicket => ({
-  ...ticket,
-  section: reverseSection(ticket.section),
-});
-const reverseTickets = (
-  ...tickets: readonly ExpressTicket[]
-): readonly ExpressTicket[] => tickets.map(reverseTicket).reverse();
-
-const kilometerFormatter = new Intl.NumberFormat(undefined, {
-  style: "unit",
-  unit: "kilometer",
-  minimumFractionDigits: 1,
-});
-
-const jpyNameFormatter = new Intl.NumberFormat(undefined, {
+type Campaign =
+  | "regular"
+  | "shinkansenYear"
+  | "limited35Percent"
+  | "shinshuPreDc";
+type Tab = "detail" | "ranking";
+type RankingFacility =
+  | "ordinary"
+  | "green"
+  | "granClassNoRefreshments"
+  | "granClassWithRefreshments";
+type RankingLimit = 50 | 100 | "all";
+type OrdinaryRankingBasis = "nonReserved" | "reserved";
+type ExclusionReason =
+  | "historicalFacility"
+  | "limitedFacility"
+  | "shinshuPreDc"
+  | "invalidJourney";
+
+const yen = new Intl.NumberFormat("ja-JP", {
   style: "currency",
   currency: "JPY",
-  currencyDisplay: "name",
+  maximumFractionDigits: 0,
 });
+const integer = new Intl.NumberFormat("ja-JP");
 
-const jpySymbolFormatter = new Intl.NumberFormat(undefined, {
-  style: "currency",
-  currency: "JPY",
-  currencyDisplay: "symbol",
-});
+const versionOptions: readonly { value: DataVersion; label: string }[] = [
+  { value: "2026-03-14", label: "2026年3月14日以降" },
+  { value: "2022-03-12", label: "2022年3月12日時点" },
+];
 
-const sortSection = (
-  section: Section
-): Readonly<{ section: SortedSection; reversed: boolean }> =>
-  section.departure.index < section.arrival.index
-    ? { section: { ...section, sorted: true }, reversed: false }
-    : { section: { ...reverseSection(section), sorted: true }, reversed: true };
-
-const Result: React.FC<{
-  line: Line;
-  section: Section;
-  highSpeed: Section | undefined;
-  longestHighSpeedSection: Section | undefined;
-  rankedFares: Record<
-    "nonReservedOrStandingOnly" | "reserved" | "reservedHighSpeed",
-    readonly {
-      readonly section: Section;
-      readonly rank: number;
-    }[]
-  >;
-  season: Season;
-  getPoints(distance: number): number;
-}> = ({
-  line,
-  section,
-  highSpeed,
-  longestHighSpeedSection,
-  rankedFares,
-  season,
-  getPoints,
-}) => {
-  const { section: sortedSection, reversed } = sortSection(section);
-  const sortedHighSpeed = highSpeed && sortSection(highSpeed).section;
-
-  const { distance, points, ...others } = getFares({
-    line,
-    section: sortedSection,
-    highSpeed: sortedHighSpeed,
-    season,
-    getPoints,
-  });
-
-  const { nonReservedOrStandingOnly, reservedHighSpeed } = others;
-
-  const totalFares = (
-    ["nonReservedOrStandingOnly", "reserved", "reservedHighSpeed"] as const
-  ).flatMap((key: keyof typeof others) => {
-    const totalFare: TotalFare | undefined = others[key];
-    return totalFare
-      ? [
-          {
-            ...totalFare,
-            key,
-            ...(reversed
-              ? { expressTickets: reverseTickets(...totalFare.expressTickets) }
-              : {}),
-          },
-        ]
-      : [];
-  });
-
-  const cells0 = totalFares.map(({ expressTickets, key }) => (
-    <th scope="col" key={key}>
-      {expressTickets.some(({ highSpeed }) => highSpeed) && (
-        <>
-          {highSpeedTrains.get(line)!}号<br />
-        </>
-      )}
-      <SeatsLabel tickets={expressTickets} />
-    </th>
-  ));
-
-  return (
-    <>
-      <h2 className="h4 mt-4">営業キロ</h2>
-      <p>{kilometerFormatter.format(distance)}</p>
-      <h2 className="h4 mt-4">所定の運賃・特急料金</h2>
-      <Table bordered>
-        <thead>
-          <tr>
-            <td></td>
-            {cells0}
-          </tr>
-        </thead>
-        <tbody>
-          <tr>
-            <th scope="row">運賃</th>
-            {totalFares.map(({ basicFare, key }) => (
-              <td key={key}>{jpyNameFormatter.format(basicFare)}</td>
-            ))}
-          </tr>
-          <tr>
-            <th scope="row">特急料金</th>
-            {totalFares.map(({ expressTickets, key }) => (
-              <td key={key}>
-                <ExpressFaresLabel tickets={expressTickets} />
-              </td>
-            ))}
-          </tr>
-          {totalFares.some(({ discount }) => discount !== undefined) ? (
-            <tr>
-              <th scope="row">割引</th>
-              {totalFares.map(({ discount, key }) => (
-                <td key={key}>
-                  {discount !== undefined
-                    ? jpyNameFormatter.format(discount)
-                    : undefined}
-                </td>
-              ))}
-            </tr>
-          ) : undefined}
-        </tbody>
-        <tfoot>
-          <tr>
-            <th scope="row">計</th>
-            {totalFares.map(({ total, key }) => (
-              <td key={key}>{jpyNameFormatter.format(total)}</td>
-            ))}
-          </tr>
-        </tfoot>
-      </Table>
-      <details className="mb-3">
-        <summary className="h6">きっぷの種類</summary>
-        <p>
-          利用可能な種類のきっぷのうち、最も低廉なものの運賃・特急料金を表示しています。
-        </p>
-        <dl>
-          {totalFares.map(({ expressTickets, types, key }) => (
-            <Row key={key}>
-              <Col as="dt" xs="auto">
-                {expressTickets.some(({ highSpeed }) => highSpeed) && (
-                  <>{highSpeedTrains.get(line)!}号 </>
-                )}
-                <SeatsLabel tickets={expressTickets} />
-              </Col>
-              <Col as="dd" xs>
-                {types.map(({ name, url }, index) => (
-                  <Fragment key={name}>
-                    {url !== undefined ? (
-                      <a href={url.href} target="_blank" rel="noreferrer">
-                        {name}
-                      </a>
-                    ) : (
-                      name
-                    )}
-                    {types.length - 1 !== index ? "・" : ""}
-                  </Fragment>
-                ))}
-                {types.length > 1 ? "のいずれか" : ""}
-              </Col>
-            </Row>
-          ))}
-        </dl>
-      </details>
-      <h2 className="h4 mt-4">JRE POINT特典チケット</h2>
-      {points !== undefined ? (
-        <>
-          <h3 className="h6">交換ポイント</h3>
-          <p>{points.toLocaleString()}ポイント</p>
-          <h3 className="h6">レート</h3>
-          <p>
-            所定額（運賃・特急料金・割引の合計）のそれぞれを交換ポイントで割った値です。
-            レートが高ければ高いほど、割引なしのきっぷに比べてお得であるといえます。
-          </p>
-          <Table bordered>
-            <thead>
-              <tr>
-                <td></td>
-                {cells0}
-              </tr>
-            </thead>
-            <tbody>
-              <tr>
-                <th scope="row">レート</th>
-                {totalFares.map(({ total, key }) => (
-                  <td key={key}>{(total / points).toFixed(2)}</td>
-                ))}
-              </tr>
-              <tr>
-                <th scope="row">順位</th>
-                {nonReservedOrStandingOnly && (
-                  <td>
-                    {rankedFares.nonReservedOrStandingOnly.find(({ section }) =>
-                      isEquivalent(section, sortedSection)
-                    )!.rank + 1}
-                    位
-                  </td>
-                )}
-                <td>
-                  {nonReservedOrStandingOnly === undefined
-                    ? `${
-                        rankedFares.nonReservedOrStandingOnly.find(
-                          ({ section }) => isEquivalent(section, sortedSection)
-                        )!.rank + 1
-                      }位または`
-                    : undefined}
-                  {rankedFares.reserved.find(({ section }) =>
-                    isEquivalent(section, sortedSection)
-                  )!.rank + 1}
-                  位
-                  {reservedHighSpeed === undefined
-                    ? `または${
-                        rankedFares.reservedHighSpeed.find(({ section }) =>
-                          isEquivalent(section, sortedSection)
-                        )!.rank + 1
-                      }位`
-                    : undefined}
-                </td>
-                {reservedHighSpeed && (
-                  <td>
-                    {highSpeed &&
-                    longestHighSpeedSection &&
-                    isEquivalent(highSpeed, longestHighSpeedSection) ? (
-                      <>
-                        {rankedFares.reservedHighSpeed.find(({ section }) =>
-                          isEquivalent(section, sortedSection)
-                        )!.rank + 1}
-                        位
-                      </>
-                    ) : undefined}
-                  </td>
-                )}
-              </tr>
-            </tbody>
-          </Table>
-        </>
-      ) : (
-        <Alert variant="warning">指定した区間では利用できません。</Alert>
-      )}
-    </>
-  );
+const facilityLabels: Readonly<Record<Facility, string>> = {
+  ordinary: "普通車指定席",
+  green: "グリーン車",
+  granClassNoRefreshments: "グランクラス（飲料・軽食なし）",
+  granClassWithRefreshments: "グランクラス（飲料・軽食あり）",
 };
 
-const ContextAwareItem: React.FC<{
-  eventKey: string;
-  line: Line;
-  section: Section;
-  highSpeed: Section;
-  longestHighSpeedSection: Section;
-  onDepartureChange: (station: Station) => void;
-  onArrivalChange: (station: Station) => void;
-  onReset: () => void;
-}> = ({
-  eventKey,
-  line,
-  section,
-  highSpeed,
-  longestHighSpeedSection,
-  onDepartureChange,
-  onArrivalChange,
-  onReset,
-}) => {
-  const { activeEventKey } = useContext(AccordionContext);
-
-  const isCurrentEventKey = activeEventKey === eventKey;
-  const stations1 = (
-    section.departure.index < section.arrival.index
-      ? line.slice(section.departure.index, section.arrival.index + 1)
-      : line.slice(section.arrival.index, section.departure.index + 1)
-  ).filter((station) => isHighSpeedAvailableStation(line, station));
-  const items = stations1.map((station) => ({
-    station,
-    disabled:
-      station.index <= line0.findIndex((station) => station.name === "大宮") ||
-      station.index >= line0.findIndex((station) => station.name === "盛岡"),
-  }));
-
-  const train = highSpeedTrains.get(line)!;
-
-  return (
-    <Accordion.Item eventKey={eventKey}>
-      <Accordion.Header>
-        <div className="d-flex justify-content-between flex-grow-1 overflow-hidden me-1">
-          <span className="flex-shrink-0">{train}号の利用区間</span>
-          <Fade in={!isCurrentEventKey}>
-            <span
-              className="ms-4 overflow-hidden text-nowrap"
-              style={{ textOverflow: "ellipsis" }}
-            >
-              <b>{highSpeed.departure.name}</b>{" "}
-              <i className="bi bi-arrow-right" />{" "}
-              <b>{highSpeed.arrival.name}</b>
-            </span>
-          </Fade>
-        </div>
-      </Accordion.Header>
-      <Accordion.Body as="fieldset">
-        <Row className="gy-2 gx-3">
-          <Col>
-            <StationDropdown
-              value={highSpeed.departure}
-              onChange={onDepartureChange}
-              header={`${train}号の乗車駅`}
-              items={items.map(({ station, disabled }) => ({
-                station,
-                disabled:
-                  station === longestHighSpeedSection.arrival ||
-                  (disabled && station !== section.departure),
-                active: station === highSpeed.departure,
-              }))}
-            />
-          </Col>
-          <Col xs="auto" className="align-self-center">
-            <i className="bi bi-arrow-right" />
-          </Col>
-          <Col>
-            <StationDropdown
-              value={highSpeed.arrival}
-              onChange={onArrivalChange}
-              header={`${train}号の降車駅`}
-              items={items.map(({ station, disabled }) => ({
-                station,
-                disabled:
-                  station === longestHighSpeedSection.departure ||
-                  (disabled && station !== section.arrival),
-                active: station === highSpeed.arrival,
-              }))}
-            />
-          </Col>
-        </Row>
-        <div className="d-grid mt-3">
-          <ToggleButton
-            variant="outline-secondary"
-            onChange={onReset}
-            type="checkbox"
-            id="toggle-check"
-            value="1"
-            disabled={
-              highSpeed.departure === longestHighSpeedSection?.departure &&
-              highSpeed.arrival === longestHighSpeedSection?.arrival
-            }
-            checked={
-              highSpeed.departure === longestHighSpeedSection?.departure &&
-              highSpeed.arrival === longestHighSpeedSection?.arrival
-            }
-          >
-            デフォルト
-          </ToggleButton>
-        </div>
-      </Accordion.Body>
-    </Accordion.Item>
-  );
+const seasonLabels: Readonly<Record<Season, string>> = {
+  閑散期: "閑散期",
+  通常期: "通常期",
+  繁忙期: "繁忙期",
+  最繁忙期: "最繁忙期",
 };
 
-/**
- * はやぶさ号やこまち号が利用可能な駅かどうか調べる
- * @param line
- * @param station
- * @returns はやぶさ号やこまち号が利用可能な駅ならtrue
- */
-const isHighSpeedAvailableStation = (line: Line, station: Station) =>
-  station.index <= line.findIndex((station) => station.name === "大宮") ||
-  ((line === line0 || line === line1) &&
-    station.index >= line0.findIndex((station) => station.name === "仙台"));
+const distanceBetween = (a: Station, b: Station) =>
+  Math.round(Math.abs(b.distance - a.distance) * 10) / 10;
 
-/**
- * 全乗車区間のうち、はやぶさ号やこまち号が利用可能な最長区間を求める。
- * @param line
- * @param section 全乗車区間
- * @returns はやぶさ号やこまち号が利用可能な最長区間。 `undefined` の場合は利用可能な区間がない。
- */
-const getLongestHighSpeedSection = (
+const getCurrentBasicFare = (
   line: Line,
-  section: SortedSection
-): SortedSection | undefined => {
-  const { departure, arrival } = section;
-  const highSpeedAvailableStations = line
-    .slice(departure.index, arrival.index + 1)
-    .filter((station) => isHighSpeedAvailableStation(line, station));
+  section: SortedSection,
+): number => {
+  const distanceKm = distanceBetween(section.departure, section.arrival);
+  if (line !== legacy2022Engine.line1) {
+    return get2026EastBasicFare(distanceKm);
+  }
 
-  const highSpeedDeparture = highSpeedAvailableStations.find(
-    ({ index }) => index >= departure.index
-  );
-  const highSpeedArrival = highSpeedAvailableStations
-    .filter(({ index }) => index <= arrival.index)
-    .slice(-1)[0];
+  const morioka = line.find(({ name }) => name === "盛岡")!;
+  const omagari = line.find(({ name }) => name === "大曲")!;
+  if (
+    morioka.index <= section.departure.index &&
+    section.arrival.index <= omagari.index
+  ) {
+    return get2026EastBasicFare(distanceKm, distanceKm);
+  }
 
-  return highSpeedDeparture &&
-    highSpeedArrival &&
-    highSpeedDeparture !== highSpeedArrival
-    ? { departure: highSpeedDeparture, arrival: highSpeedArrival, sorted: true }
-    : undefined;
+  const localStart =
+    section.departure.index < morioka.index ? morioka : section.departure;
+  const localEnd = section.arrival.index > omagari.index ? omagari : section.arrival;
+  const localKm =
+    localStart.index < localEnd.index ? distanceBetween(localStart, localEnd) : 0;
+  return get2026EastBasicFare(distanceKm, localKm);
 };
 
-/*
-AB -O-- --S- -M- -A
+const fareTickets = (fare: TotalFare): readonly ExpressTicket[] =>
+  fare.expressTickets;
 
-T  xxxx xxoo ooo oo 
-|   xxx xxoo ooo oo
-O    xx xxoo ooo oo
-|     x xxxo ooo oo
-|       xxxo ooo oo
-
-|        xxo ooo oo
-|         xo ooo oo
-S          o ooo oo
-|            ooo oo
-
-|             oo oo
-M              x xx
-|                xx
-
-|                 x
-*/
-/**
- * はやぶさ号やこまち号が利用可能な区間かどうか調べる。
- * @param line
- * @param section
- * @returns はやぶさ号やこまち号が利用可能な区間ならtrue
- */
-const isHighSpeedAvailableSection = (line: Line, section: SortedSection) =>
-  (line === line0 || line === line1) &&
-  ((section.departure.index <=
-    line0.findIndex((station) => station.name === "大宮") &&
-    section.arrival.index >=
-      line0.findIndex((station) => station.name === "仙台")) ||
-    (section.departure.index <
-      line0.findIndex((station) => station.name === "盛岡") &&
-      section.arrival.index >
-        line0.findIndex((station) => station.name === "仙台")));
-
-interface State {
-  readonly group: LineGroup;
+interface QuoteInput {
+  readonly version: DataVersion;
+  readonly campaign: Campaign;
   readonly line: Line;
-  readonly section: Section;
-  readonly highSpeed:
-    | {
-        readonly departure: Station | undefined;
-        readonly arrival: Station | undefined;
-      }
-    | undefined;
+  readonly section: SortedSection;
+  readonly highSpeed?: SortedSection;
+  readonly green?: Interval;
+  readonly granClass?: Interval;
+  readonly granClassWithRefreshments?: Interval;
+  readonly season: Season;
 }
 
-type Action = Readonly<
-  | {
-      type: "setGroup";
-      payload: LineGroup;
-    }
-  | {
-      type: "setDeparture";
-      payload: Station;
-    }
-  | {
-      type: "setArrival";
-      payload: Station;
-    }
-  | {
-      type: "setHighSpeedDeparture";
-      payload: Station;
-    }
-  | {
-      type: "setHighSpeedArrival";
-      payload: Station;
-    }
-  | {
-      type: "resetHighSpeed";
-      payload?: undefined;
-    }
->;
-
-interface Section {
-  departure: Station;
-  arrival: Station;
+interface Quote {
+  readonly distanceKm: number;
+  readonly points?: number | undefined;
+  readonly paperFare?: number | undefined;
+  readonly nonReservedFare?: number | undefined;
+  readonly basicFare?: number | undefined;
+  readonly expressFare?: number | undefined;
+  readonly specialVehicleFare?: number | undefined;
+  readonly facility: Facility;
+  readonly exclusionReason?: ExclusionReason | undefined;
 }
 
-/**
- * `arrival` が `departure` よりも終点に近いことが保証された `Section`
- */
-interface SortedSection extends Section {
-  sorted: true;
-}
+const requestedFacility = (journey: JourneySelection): Facility =>
+  journey.granClassWithRefreshments
+    ? "granClassWithRefreshments"
+    : journey.granClass
+      ? "granClassNoRefreshments"
+      : journey.green
+        ? "green"
+        : "ordinary";
 
-const reducer: Reducer<State, Action> = (state, { type, payload }): State => {
-  switch (type) {
-    case "setGroup": {
-      const { departure, arrival } = state.section;
-      const line = payload.lines[0]!;
-      const firstOfLine = line[0]!;
-      const lastOfLine = line.slice(-1)[0]!;
-
-      const section: Section = line.includes(departure)
-        ? line.includes(arrival)
-          ? state.section
-          : {
-              departure,
-              arrival: lastOfLine === departure ? firstOfLine : lastOfLine,
-            }
-        : line.includes(arrival)
-        ? {
-            departure: firstOfLine === arrival ? lastOfLine : firstOfLine,
-            arrival,
-          }
-        : { departure: firstOfLine, arrival: lastOfLine };
-
-      return {
-        ...state,
-        group: payload,
-        line,
-        section,
-        highSpeed: undefined,
-      };
-    }
-
-    case "setDeparture": {
-      const arrival = state.section.arrival;
-      const line = state.line.includes(payload)
-        ? state.line
-        : state.group.lines.find((line) => line.includes(payload))!;
-      const firstOfLine = line[0]!;
-      const lastOfLine = line.slice(-1)[0]!;
-
-      const section = {
-        departure: payload,
-        arrival: line.includes(arrival)
-          ? arrival
-          : lastOfLine === payload
-          ? firstOfLine
-          : lastOfLine,
-      };
-
-      return {
-        ...state,
-        line,
-        section,
-        highSpeed: undefined,
-      };
-    }
-
-    case "setArrival": {
-      const departure = state.section.departure;
-      const line = state.line.includes(payload)
-        ? state.line
-        : state.group.lines.find((line) => line.includes(payload))!;
-      const firstOfLine = line[0]!;
-      const lastOfLine = line.slice(-1)[0]!;
-
-      const section = {
-        departure: line.includes(departure)
-          ? departure
-          : firstOfLine === payload
-          ? lastOfLine
-          : firstOfLine,
-        arrival: payload,
-      };
-
-      return {
-        ...state,
-        line,
-        section,
-        highSpeed: undefined,
-      };
-    }
-
-    case "setHighSpeedDeparture":
-      return {
-        ...state,
-        highSpeed: {
-          departure: payload,
-          arrival:
-            state.highSpeed &&
-            state.highSpeed.arrival &&
-            payload.index < state.highSpeed.arrival.index
-              ? state.highSpeed.arrival
-              : undefined,
-        },
-      };
-
-    case "setHighSpeedArrival":
-      return {
-        ...state,
-        highSpeed: {
-          departure:
-            state.highSpeed &&
-            state.highSpeed.departure &&
-            state.highSpeed.departure.index < payload.index
-              ? state.highSpeed.departure
-              : undefined,
-          arrival: payload,
-        },
-      };
-
-    case "resetHighSpeed":
-      return {
-        ...state,
-        highSpeed: undefined,
-      };
-  }
-};
-
-const init = (): State => {
-  const group = lineGroups.get("東北新幹線")!;
-  const line = group.lines[0]!;
-  const section: Section = {
-    departure: line[0]!,
-    arrival: line.slice(-1)[0]!,
+export const createQuote = ({
+  version,
+  campaign,
+  line,
+  section,
+  highSpeed,
+  green,
+  granClass,
+  granClassWithRefreshments,
+  season,
+}: QuoteInput): Quote => {
+  const distanceKm = distanceBetween(section.departure, section.arrival);
+  const journey: JourneySelection = {
+    origin: section.departure.index,
+    destination: section.arrival.index,
+    ...(highSpeed
+      ? { highSpeed: { start: highSpeed.departure.index, end: highSpeed.arrival.index } }
+      : {}),
+    ...(green ? { green } : {}),
+    ...(granClass ? { granClass } : {}),
+    ...(granClassWithRefreshments ? { granClassWithRefreshments } : {}),
   };
-  return {
-    group,
+  let facility: Facility;
+  try {
+    facility = highestFacility(journey);
+  } catch (error) {
+    if (!(error instanceof RangeError)) throw error;
+    return {
+      distanceKm,
+      facility: requestedFacility(journey),
+      exclusionReason: "invalidJourney",
+    };
+  }
+
+  const legacy = legacy2022Engine.getFares({
     line,
     section,
-    highSpeed: undefined,
+    highSpeed,
+    season,
+    getPoints: (distance) => get2022Points(distance),
+  });
+  const selectedLegacyFare = highSpeed
+    ? (legacy.reservedHighSpeed ?? legacy.reserved)
+    : legacy.reserved;
+
+  if (version === "2022-03-12") {
+    if (facility !== "ordinary") {
+      return {
+        distanceKm,
+        facility,
+        exclusionReason: "historicalFacility",
+      };
+    }
+    const points = get2022Points(
+      distanceKm,
+      campaign === "shinkansenYear" ? "shinkansenYear" : "regular",
+    );
+    const cheapestFare = selectedLegacyFare.total;
+    const paperFare = cheapestFare - (selectedLegacyFare.discount ?? 0);
+    const nonReservedFare = legacy.nonReservedOrStandingOnly
+      ? legacy.nonReservedOrStandingOnly.total -
+        (legacy.nonReservedOrStandingOnly.discount ?? 0)
+      : undefined;
+    return {
+      distanceKm,
+      facility,
+      points,
+      paperFare,
+      ...(nonReservedFare !== undefined ? { nonReservedFare } : {}),
+      basicFare: selectedLegacyFare.basicFare,
+      expressFare: fareTickets(selectedLegacyFare).reduce(
+        (total, ticket) => total + ticket.fare,
+        0,
+      ),
+      specialVehicleFare: 0,
+    };
+  }
+
+  const points =
+    campaign === "shinshuPreDc"
+      ? get2026ShinshuPreDcPoints(
+          section.departure.name,
+          section.arrival.name,
+          facility,
+        )
+      : get2026JourneyPoints(
+          distanceKm,
+          journey,
+          campaign === "limited35Percent" ? "limited35Percent" : "regular",
+        );
+  if (points === undefined) {
+    return {
+      distanceKm,
+      facility,
+      exclusionReason:
+        campaign === "shinshuPreDc"
+          ? "shinshuPreDc"
+          : "limitedFacility",
+    };
+  }
+  const basicFare = getCurrentBasicFare(line, section);
+  const expressFare = get2026ExpressFare(
+    fareTickets(selectedLegacyFare),
+    green !== undefined,
+  );
+  const intervalDistance = (interval: Interval) => {
+    const start = line[interval.start];
+    const end = line[interval.end];
+    if (!start || !end) return undefined;
+    const distanceKm = distanceBetween(start, end);
+    return distanceKm > 0 ? distanceKm : undefined;
+  };
+  const greenKm = green ? intervalDistance(green) : undefined;
+  const granClassKm = granClass ? intervalDistance(granClass) : undefined;
+  const granClassWithRefreshmentsKm = granClassWithRefreshments
+    ? intervalDistance(granClassWithRefreshments)
+    : undefined;
+  if (
+    (green && greenKm === undefined) ||
+    (granClass && granClassKm === undefined) ||
+    (granClassWithRefreshments && granClassWithRefreshmentsKm === undefined)
+  ) {
+    return { distanceKm, facility, exclusionReason: "invalidJourney" };
+  }
+  const specialVehicleFare = greenKm
+    ? get2026SpecialVehicleFare({
+        greenKm,
+        ...(granClassKm !== undefined ? { granClassKm } : {}),
+        ...(granClassWithRefreshmentsKm !== undefined
+          ? { granClassWithRefreshmentsKm }
+          : {}),
+      })
+    : 0;
+  const paperFare = basicFare + expressFare + specialVehicleFare;
+  const nonReservedFare =
+    facility === "ordinary" && legacy.nonReservedOrStandingOnly
+      ? basicFare +
+        get2026ExpressFare(
+          fareTickets(legacy.nonReservedOrStandingOnly),
+          false,
+        )
+      : undefined;
+
+  return {
+    distanceKm,
+    facility,
+    points,
+    paperFare,
+    ...(nonReservedFare !== undefined ? { nonReservedFare } : {}),
+    basicFare,
+    expressFare,
+    specialVehicleFare,
   };
 };
 
-const Home: React.FC<{
-  season: Season;
-  onSeasonChange: (season: Season) => void;
-  pointTicketType: PointTicketType;
-  onPointTicketTypeChange: (type: PointTicketType) => void;
-  rankedFares: Record<
-    "nonReservedOrStandingOnly" | "reserved" | "reservedHighSpeed",
-    readonly {
-      readonly section: Section;
-      readonly rank: number;
-    }[]
-  >;
-}> = ({
-  season,
-  pointTicketType,
-  onSeasonChange,
-  onPointTicketTypeChange,
-  rankedFares,
-}) => {
-  const [state, dispatch] = useReducer(reducer, undefined, init);
-
-  const { group, line, section } = state;
-
-  const { section: sortedSection, reversed } = useMemo(
-    () => sortSection(section),
-    [section]
-  );
-
-  const longestHighSpeedSection = useMemo(() => {
-    const longestHighSpeedSection = getLongestHighSpeedSection(
-      line,
-      sortedSection
-    );
-    return reversed
-      ? longestHighSpeedSection && reverseSection(longestHighSpeedSection)
-      : longestHighSpeedSection;
-  }, [line, reversed, sortedSection]);
-
-  const highSpeedAvailable = useMemo(
-    () => isHighSpeedAvailableSection(line, sortedSection),
-    [line, sortedSection]
-  );
-
-  const highSpeed: Section | undefined =
-    longestHighSpeedSection && highSpeedAvailable
-      ? {
-          departure:
-            state.highSpeed?.departure ?? longestHighSpeedSection.departure,
-          arrival: state.highSpeed?.arrival ?? longestHighSpeedSection.arrival,
-        }
-      : undefined;
-
-  const { departure, arrival } = section;
-
-  return (
-    <main>
-      <h1>区間を指定して調べる</h1>
-      <p>
-        えきねっとで交換できる
-        <a
-          href="https://www.eki-net.com/top/point/guide/tokuten_section.html#headerM2_01"
-          target="_blank"
-          rel="noreferrer"
-        >
-          JRE POINT特典チケット
-        </a>
-        のレート（1ポイントあたり何円相当か）を、区間ごとに調べられます。
-        自由席と指定席など、それぞれの所定額に対するレートと順位が同時に表示されるため、さまざまな観点で検討できます。
-      </p>
-      <Card body className="my-3" as="fieldset">
-        <FloatingLabel controlId="floatingSelect" label="路線" className="mb-3">
-          <Form.Select
-            aria-label="Floating label select example"
-            onChange={(e) =>
-              dispatch({
-                type: "setGroup",
-                payload: lineGroups.get(e.currentTarget.value)!,
-              })
-            }
-          >
-            <option>東北新幹線</option>
-            <option>秋田新幹線</option>
-            <option>山形新幹線</option>
-            <option>上越新幹線</option>
-            <option>北陸新幹線</option>
-          </Form.Select>
-        </FloatingLabel>
-        <Row className="gy-2 gx-3">
-          <Col>
-            <StationDropdown
-              value={departure}
-              onChange={(station) =>
-                dispatch({ type: "setDeparture", payload: station })
-              }
-              header="発駅"
-              items={[...new Set(group.lines.flat())]
-                .sort((a, b) => a.distance - b.distance)
-                .map((station) => ({
-                  station,
-                  disabled: station === arrival,
-                  active: station === departure,
-                }))}
-            />
-          </Col>
-          <Col xs="auto" className="align-self-center">
-            <i className="bi bi-arrow-right" />
-          </Col>
-          <Col>
-            <StationDropdown
-              value={arrival}
-              onChange={(station) =>
-                dispatch({ type: "setArrival", payload: station })
-              }
-              header="着駅"
-              items={[...new Set(group.lines.flat())]
-                .sort((a, b) => a.distance - b.distance)
-                .map((station) => ({
-                  station,
-                  disabled: station === departure,
-                  active: station === arrival,
-                }))}
-            />
-          </Col>
-        </Row>
-      </Card>
-      {highSpeed && longestHighSpeedSection ? (
-        <Accordion className="mb-3">
-          <ContextAwareItem
-            eventKey="0"
-            line={line}
-            section={section}
-            highSpeed={highSpeed}
-            longestHighSpeedSection={longestHighSpeedSection}
-            onDepartureChange={(station: Station) =>
-              dispatch({ type: "setHighSpeedDeparture", payload: station })
-            }
-            onArrivalChange={(station: Station) =>
-              dispatch({ type: "setHighSpeedArrival", payload: station })
-            }
-            onReset={() => dispatch({ type: "resetHighSpeed" })}
-          />
-        </Accordion>
-      ) : undefined}
-      <Row className="gy-2 mb-3">
-        <Col md>
-          <SeasonSelect
-            season={season}
-            onChange={(season) => onSeasonChange(season)}
-          />
-        </Col>
-        <Col md>
-          <PointTicketTypeSelect
-            type={pointTicketType}
-            onChange={(type) => onPointTicketTypeChange(type)}
-          />
-        </Col>
-      </Row>
-      <Result
-        line={line}
-        section={section}
-        highSpeed={highSpeed}
-        longestHighSpeedSection={longestHighSpeedSection}
-        rankedFares={rankedFares}
-        season={season}
-        getPoints={pointTicketType.getPoints}
-      />
-    </main>
-  );
+export const intervalWithin = (
+  enabled: boolean,
+  start: number,
+  end: number,
+  outerStart: number,
+  outerEnd: number,
+): Interval | undefined => {
+  if (!enabled || !(outerStart < outerEnd)) return undefined;
+  const clampedStart = Math.max(outerStart, Math.min(start, outerEnd - 1));
+  const clampedEnd = Math.min(outerEnd, Math.max(end, outerStart + 1));
+  return clampedStart < clampedEnd
+    ? { start: clampedStart, end: clampedEnd }
+    : { start: outerStart, end: outerEnd };
 };
 
-type Ranked<T> = { value: T; rank: number };
-const rank = <T,>(
-  array: readonly T[],
-  callbackfn: (t: T) => number
-): readonly Ranked<T>[] =>
-  [...array]
-    .sort((a, b) => callbackfn(b) - callbackfn(a))
-    .reduce<{
-      readonly last?: Ranked<T>;
-      readonly array: readonly Ranked<T>[];
-    }>(
-      (previous, current, index) => {
-        const last = {
-          value: current,
-          rank:
-            previous.last === undefined ||
-            callbackfn(previous.last.value) !== callbackfn(current)
-              ? index
-              : previous.last.rank,
-        };
-        return { last, array: [...previous.array, last] };
-      },
-      { array: [] }
-    ).array;
+const granClassLastIndex = (line: Line) => {
+  if (line === legacy2022Engine.line1) {
+    return line.find(({ name }) => name === "盛岡")!.index;
+  }
+  if (line === legacy2022Engine.line2) {
+    return line.find(({ name }) => name === "福島")!.index;
+  }
+  return line.length - 1;
+};
 
-const Ranking: React.FC<{
-  rankedFares: Record<
-    "nonReservedOrStandingOnly" | "reserved" | "reservedHighSpeed",
-    readonly {
-      readonly section: Section;
-      readonly rank: number;
-      readonly points: number;
-      readonly distance: number;
-      readonly nonReservedOrStandingOnly:
-        | (TotalFare & { readonly rate: number })
-        | undefined;
-      readonly reserved: TotalFare & { readonly rate: number };
-      readonly reservedHighSpeed:
-        | (TotalFare & { readonly rate: number })
-        | undefined;
-    }[]
-  >;
-  season: Season;
-  onSeasonChange: (season: Season) => void;
-  pointTicketType: PointTicketType;
-  onPointTicketTypeChange: (type: PointTicketType) => void;
-}> = ({
-  rankedFares,
-  season,
-  onSeasonChange,
-  pointTicketType,
-  onPointTicketTypeChange,
+const canonicalFacilitySections = (
+  line: Line,
+  section: SortedSection,
+  facility: RankingFacility,
+) => {
+  if (facility === "ordinary") return {};
+
+  const green: Interval = {
+    start: section.departure.index,
+    end: section.arrival.index,
+  };
+  if (facility === "green") return { green };
+
+  const granClassEnd = Math.min(section.arrival.index, granClassLastIndex(line));
+  if (section.departure.index >= granClassEnd) return undefined;
+  const granClass: Interval = {
+    start: section.departure.index,
+    end: granClassEnd,
+  };
+  return {
+    green,
+    granClass,
+    ...(facility === "granClassWithRefreshments"
+      ? { granClassWithRefreshments: granClass }
+      : {}),
+  };
+};
+
+const rate = (fare: number | undefined, points: number | undefined) =>
+  fare !== undefined && points !== undefined ? fare / points : undefined;
+
+const RateCard = ({
+  label,
+  fare,
+  points,
+}: {
+  label: string;
+  fare?: number | undefined;
+  points?: number | undefined;
+}) => (
+  <article className="rate-card">
+    <span>{label}</span>
+    <strong>{fare === undefined ? "—" : yen.format(fare)}</strong>
+    <small>{rate(fare, points)?.toFixed(2) ?? "—"} 円/pt</small>
+  </article>
+);
+
+export const rangeAfterStartChange = (
+  stations: readonly Station[],
+  end: number,
+  start: number,
+) => {
+  const position = stations.findIndex((station) => station.index === start);
+  const next = stations[position + 1];
+  return { start, end: start >= end && next ? next.index : end };
+};
+
+export const rangeAfterEndChange = (
+  stations: readonly Station[],
+  start: number,
+  end: number,
+) => {
+  const position = stations.findIndex((station) => station.index === end);
+  const previous = stations[position - 1];
+  return { start: end <= start && previous ? previous.index : start, end };
+};
+
+const RangeSelect = ({
+  label,
+  stations,
+  start,
+  end,
+  onStart,
+  onEnd,
+}: {
+  label: string;
+  stations: readonly Station[];
+  start: number;
+  end: number;
+  onStart: (value: number) => void;
+  onEnd: (value: number) => void;
 }) => {
-  const [seat, setSeat] = useState<
-    "nonReservedOrStandingOnly" | "reserved" | "reservedHighSpeed"
-  >("nonReservedOrStandingOnly");
-  const [small, setSmall] = useState(false);
+  const selectStart = (value: number) => {
+    const range = rangeAfterStartChange(stations, end, value);
+    if (range.end !== end) onEnd(range.end);
+    onStart(range.start);
+  };
+  const selectEnd = (value: number) => {
+    const range = rangeAfterEndChange(stations, start, value);
+    if (range.start !== start) onStart(range.start);
+    onEnd(range.end);
+  };
 
   return (
-    <main>
-      <h1>ランキング</h1>
-      <p>
-        えきねっとで交換できる
-        <a
-          href="https://www.eki-net.com/top/point/guide/tokuten_section.html#headerM2_01"
-          target="_blank"
-          rel="noreferrer"
-        >
-          JRE POINT特典チケット
-        </a>
-        のレート（1ポイントあたり何円相当か）のランキングです。
-        割引なしのきっぷと比べたとき、どの区間でJRE POINT
-        特典チケットを利用するのがお得かを調べられます。
-      </p>
-      <Form.Group className="mb-3" controlId="SmallTableCheckbox">
-        <Form.Check
-          type="checkbox"
-          label="小さく表示する"
-          checked={small}
-          onChange={(e) => setSmall(e.currentTarget.checked)}
-        />
-      </Form.Group>
-      <Row className="gy-2 mb-3">
-        <Col xl>
-          <FloatingLabel label="次の所定額に対するレートで並び替える">
-            <Form.Select
-              value={seat}
-              onChange={(e) =>
-                setSeat(
-                  e.currentTarget.value as
-                    | "nonReservedOrStandingOnly"
-                    | "reserved"
-                    | "reservedHighSpeed"
-                )
-              }
-            >
-              <option value="nonReservedOrStandingOnly">自由席・立席</option>
-              <option value="reserved">指定席</option>
-              <option value="reservedHighSpeed">
-                はやぶさ号・こまち号 指定席
-              </option>
-            </Form.Select>
-          </FloatingLabel>
-        </Col>
-        <Col xl>
-          <SeasonSelect
-            season={season}
-            onChange={(season) => onSeasonChange(season)}
-          />
-        </Col>
-        <Col xl>
-          <PointTicketTypeSelect
-            type={pointTicketType}
-            onChange={(type) => onPointTicketTypeChange(type)}
-          />
-        </Col>
-      </Row>
-      <Table
-        striped
-        bordered
-        responsive
-        {...(small ? { size: "sm" } : {})}
-        className="text-nowrap"
+    <div className="range-select">
+      <span>{label}</span>
+      <select
+        aria-label={`${label} 始点`}
+        value={start}
+        onChange={(event) => selectStart(Number(event.target.value))}
       >
-        <thead>
-          <tr>
-            <th scope="col" rowSpan={2}>
-              #
-            </th>
-            <th scope="col" colSpan={2} rowSpan={2}>
-              区間
-            </th>
-            <th scope="col" rowSpan={2}>
-              営業キロ
-            </th>
-            <th scope="col" rowSpan={2}>
-              交換ポイント
-            </th>
-            <th scope="col" colSpan={2}>
-              自由席・立席
-            </th>
-            <th scope="col" colSpan={2}>
-              指定席
-            </th>
-            <th scope="col" colSpan={2}>
-              はやぶさ号・こまち号
-              <br />
-              指定席
-            </th>
-          </tr>
-          <tr>
-            <th scope="col">所定額</th>
-            <th scope="col">レート</th>
-            <th scope="col">所定額</th>
-            <th scope="col">レート</th>
-            <th scope="col">所定額</th>
-            <th scope="col">レート</th>
-          </tr>
-        </thead>
-        <tbody>
-          {rankedFares[seat].map(
-            ({
-              rank,
-              section,
-              distance,
-              points,
-              nonReservedOrStandingOnly,
-              reserved,
-              reservedHighSpeed,
-            }) => (
-              <tr key={`${section.departure.name}-${section.arrival.name}`}>
-                <th scope="row">{rank + 1}</th>
-                <th scope="row">{section.departure.name}</th>
-                <th scope="row">{section.arrival.name}</th>
-                <td className="text-end">
-                  {kilometerFormatter.format(distance)}
-                </td>
-                <td className="text-end">{points.toLocaleString()}</td>
-                <td className="text-end">
-                  {nonReservedOrStandingOnly
-                    ? jpySymbolFormatter.format(nonReservedOrStandingOnly.total)
-                    : undefined}
-                </td>
-                <td className="text-end">
-                  {nonReservedOrStandingOnly ? (
-                    <strong
-                      className={
-                        seat === "nonReservedOrStandingOnly"
-                          ? "text-primary"
-                          : undefined
-                      }
-                    >
-                      {nonReservedOrStandingOnly.rate.toFixed(2)}
-                    </strong>
-                  ) : (
-                    <i className="text-muted">{reserved.rate.toFixed(2)}</i>
-                  )}
-                </td>
-                <td className="text-end">
-                  {jpySymbolFormatter.format(reserved.total)}
-                </td>
-                <td className="text-end">
-                  <strong
-                    className={seat === "reserved" ? "text-primary" : undefined}
-                  >
-                    {reserved.rate.toFixed(2)}
-                  </strong>
-                </td>
-                <td className="text-end">
-                  {reservedHighSpeed
-                    ? jpySymbolFormatter.format(reservedHighSpeed.total)
-                    : undefined}
-                </td>
-                <td className="text-end">
-                  {reservedHighSpeed ? (
-                    <strong
-                      className={
-                        seat === "reservedHighSpeed"
-                          ? "text-primary"
-                          : undefined
-                      }
-                    >
-                      {reservedHighSpeed.rate.toFixed(2)}
-                    </strong>
-                  ) : (
-                    <i className="text-muted">{reserved.rate.toFixed(2)}</i>
-                  )}
-                </td>
-              </tr>
-            )
-          )}
-        </tbody>
-      </Table>
-    </main>
+        {stations.slice(0, -1).map((station) => (
+          <option value={station.index} key={station.name}>{station.name}</option>
+        ))}
+      </select>
+      <span className="range-arrow">→</span>
+      <select
+        aria-label={`${label} 終点`}
+        value={end}
+        onChange={(event) => selectEnd(Number(event.target.value))}
+      >
+        {stations.slice(1).map((station) => (
+          <option value={station.index} key={station.name}>{station.name}</option>
+        ))}
+      </select>
+    </div>
   );
 };
 
-const average = "通常期";
-const busy = "繁忙期";
-const busiest = "最繁忙期";
-const off = "閑散期";
+const exclusionMessages: Readonly<
+  Record<ExclusionReason, { readonly title: string; readonly detail: string }>
+> = {
+  historicalFacility: {
+    title: "2022年版の対象外です",
+    detail: "2022年版は普通車指定席のみ参照できます。",
+  },
+  limitedFacility: {
+    title: "全線35%特別レートの対象外です",
+    detail: "飲料・軽食ありのグランクラスは対象外です。",
+  },
+  shinshuPreDc: {
+    title: "信州プレDCの対象外です",
+    detail:
+      "公式表に掲載された北陸新幹線の対象駅間・普通車指定席のみ計算できます。",
+  },
+  invalidJourney: {
+    title: "利用区間を計算できません",
+    detail: "各設備の始点と終点、および区間の包含関係を確認してください。",
+  },
+};
 
-const seasons = [off, average, busy, busiest] as const;
+const App = () => {
+  const [tab, setTab] = useState<Tab>("detail");
+  const [version, setVersion] = useState<DataVersion>("2026-03-14");
+  const [campaign, setCampaign] = useState<Campaign>("regular");
+  const [groupName, setGroupName] = useState("東北新幹線");
+  const group = legacy2022Engine.lineGroups.get(groupName)!;
+  const [routeIndex, setRouteIndex] = useState(0);
+  const line = group.lines[Math.min(routeIndex, group.lines.length - 1)]!;
+  const [departureIndex, setDepartureIndex] = useState(0);
+  const [arrivalIndex, setArrivalIndex] = useState(line.length - 1);
+  const [season, setSeason] = useState<Season>(legacy2022Engine.average);
+  const [highSpeedEnabled, setHighSpeedEnabled] = useState(false);
+  const [highSpeedStart, setHighSpeedStart] = useState(0);
+  const [highSpeedEnd, setHighSpeedEnd] = useState(line.length - 1);
+  const [greenEnabled, setGreenEnabled] = useState(false);
+  const [greenStart, setGreenStart] = useState(0);
+  const [greenEnd, setGreenEnd] = useState(line.length - 1);
+  const [granClassEnabled, setGranClassEnabled] = useState(false);
+  const [granClassStart, setGranClassStart] = useState(0);
+  const [granClassEnd, setGranClassEnd] = useState(line.length - 1);
+  const [refreshments, setRefreshments] = useState(false);
+  const [rankingFacility, setRankingFacility] =
+    useState<RankingFacility>("ordinary");
+  const [rankingLimit, setRankingLimit] = useState<RankingLimit>(50);
+  const [ordinaryRankingBasis, setOrdinaryRankingBasis] =
+    useState<OrdinaryRankingBasis>("nonReserved");
 
-type Season = typeof seasons[number];
+  const resetRoute = (nextGroupName: string, nextRouteIndex = 0) => {
+    const nextLine = legacy2022Engine.lineGroups.get(nextGroupName)!.lines[nextRouteIndex]!;
+    setGroupName(nextGroupName);
+    setRouteIndex(nextRouteIndex);
+    setDepartureIndex(0);
+    setArrivalIndex(nextLine.length - 1);
+    setHighSpeedStart(0);
+    setHighSpeedEnd(nextLine.length - 1);
+    setGreenStart(0);
+    setGreenEnd(nextLine.length - 1);
+    setGranClassStart(0);
+    setGranClassEnd(nextLine.length - 1);
+  };
 
-const SeasonSelect: React.FC<{
-  season: Season;
-  onChange: (season: Season) => void;
-}> = ({ season, onChange }) => (
-  <FloatingLabel label="シーズン">
-    <Form.Select
-      value={season}
-      onChange={(e) => onChange(e.currentTarget.value as Season)}
-    >
-      {seasons.map((season) => (
-        <option key={season}>{season}</option>
-      ))}
-    </Form.Select>
-  </FloatingLabel>
-);
-
-const PointTicketTypeSelect: React.FC<{
-  type: PointTicketType;
-  onChange: (type: PointTicketType) => void;
-}> = ({ type, onChange }) => (
-  <FloatingLabel label="JRE POINT特典チケット 交換ポイント">
-    <Form.Select
-      value={type.name}
-      onChange={(e) =>
-        onChange(
-          pointTicketTypes.find(({ name }) => name === e.currentTarget.value)!
-        )
-      }
-    >
-      {pointTicketTypes.map((nextPointTicketType) => (
-        <option key={nextPointTicketType.name}>
-          {nextPointTicketType.name}
-        </option>
-      ))}
-    </Form.Select>
-    {type.url && (
-      <Form.Text>
-        <a href={type.url.href} target="_blank" rel="noreferrer">
-          詳細
-        </a>
-      </Form.Text>
-    )}
-  </FloatingLabel>
-);
-
-const App: React.FC = () => {
-  const [season, setSeason] = useState<Season>(average);
-  const [pointTicketType, setPointTicketType] = useState(pointTicketTypes[0]!);
-  const [tab, setTab] = useState<"detail" | "ranking">("detail");
-
-  const faresForEachSection = useMemo(
-    () =>
-      [...lineGroups.values()]
-        .flatMap(({ lines }) => lines)
-        .flatMap((line) => {
-          const junction = junctions.get(line);
-          return line.flatMap((departure, index, stations) =>
-            stations
-              .slice(Math.max(index, junction?.index ?? 0) + 1)
-              .flatMap((arrival) => {
-                const section: SortedSection = {
-                  departure,
-                  arrival,
-                  sorted: true,
-                };
-                const highSpeed = getLongestHighSpeedSection(line, section);
-                const {
-                  points,
-                  distance,
-                  nonReservedOrStandingOnly,
-                  reserved,
-                  reservedHighSpeed,
-                } = getFares({
-                  line,
-                  section,
-                  highSpeed,
-                  season,
-                  getPoints: pointTicketType.getPoints,
-                });
-
-                return points !== undefined
-                  ? [
-                      {
-                        section,
-                        points,
-                        distance,
-                        nonReservedOrStandingOnly:
-                          nonReservedOrStandingOnly && {
-                            ...nonReservedOrStandingOnly,
-                            rate: nonReservedOrStandingOnly.total / points,
-                          },
-                        reserved: {
-                          ...reserved,
-                          rate: reserved.total / points,
-                        },
-                        reservedHighSpeed: reservedHighSpeed && {
-                          ...reservedHighSpeed,
-                          rate: reservedHighSpeed.total / points,
-                        },
-                      },
-                    ]
-                  : [];
-              })
-          );
-        }),
-    [pointTicketType.getPoints, season]
+  const sorted = legacy2022Engine.sortSection({
+    departure: line[Math.min(departureIndex, line.length - 2)]!,
+    arrival: line[Math.max(1, Math.min(arrivalIndex, line.length - 1))]!,
+  }).section;
+  const tripStart = sorted.departure.index;
+  const tripEnd = sorted.arrival.index;
+  const tripStations = line.slice(tripStart, tripEnd + 1);
+  const granClassLimit = granClassLastIndex(line);
+  const granClassAvailable = tripStart < Math.min(tripEnd, granClassLimit);
+  const highSpeed = intervalWithin(
+    highSpeedEnabled,
+    highSpeedStart,
+    highSpeedEnd,
+    tripStart,
+    tripEnd,
   );
-
-  const rankedFares = useMemo(
-    () => ({
-      nonReservedOrStandingOnly: rank(
-        faresForEachSection,
-        ({ reserved, nonReservedOrStandingOnly }) =>
-          nonReservedOrStandingOnly?.rate ?? reserved.rate
-      ).map(({ value, rank }) => ({ ...value, rank })),
-      reserved: rank(faresForEachSection, ({ reserved }) => reserved.rate).map(
-        ({ value, rank }) => ({ ...value, rank })
-      ),
-      reservedHighSpeed: rank(
-        faresForEachSection,
-        ({ reserved, reservedHighSpeed }) =>
-          reservedHighSpeed?.rate ?? reserved.rate
-      ).map(({ value, rank }) => ({ ...value, rank })),
-    }),
-    [faresForEachSection]
+  const green = intervalWithin(
+    greenEnabled,
+    greenStart,
+    greenEnd,
+    tripStart,
+    tripEnd,
   );
+  const granClassOuterStart = green?.start ?? tripStart;
+  const granClassOuterEnd = Math.min(green?.end ?? tripEnd, granClassLimit);
+  const granClassStations = line.slice(
+    granClassOuterStart,
+    granClassOuterEnd + 1,
+  );
+  const granClass = intervalWithin(
+    green !== undefined && granClassEnabled && granClassAvailable,
+    granClassStart,
+    granClassEnd,
+    granClassOuterStart,
+    granClassOuterEnd,
+  );
+  const refreshmentSection = refreshments && granClass ? granClass : undefined;
+  const selectedFacility: Facility = refreshments && granClass
+    ? "granClassWithRefreshments"
+    : granClass
+      ? "granClassNoRefreshments"
+      : green
+        ? "green"
+        : "ordinary";
+  const highSpeedSection = highSpeed
+    ? { departure: line[highSpeed.start]!, arrival: line[highSpeed.end]!, sorted: true as const }
+    : undefined;
+
+  const quote = createQuote({
+    version,
+    campaign,
+    line,
+    section: sorted,
+    ...(highSpeedSection ? { highSpeed: highSpeedSection } : {}),
+    ...(green ? { green } : {}),
+    ...(granClass ? { granClass } : {}),
+    ...(refreshmentSection
+      ? { granClassWithRefreshments: refreshmentSection }
+      : {}),
+    season,
+  });
+
+  const rankingRows = useMemo(() => {
+    if (tab !== "ranking") return [];
+    const rows = [...legacy2022Engine.lineGroups.entries()].flatMap(
+      ([rankingGroupName, rankingGroup]) =>
+        rankingGroup.lines.flatMap((rankingLine) =>
+          rankingLine.flatMap((departure, departureOffset) =>
+            rankingLine.slice(departureOffset + 1).flatMap((arrival) => {
+              const rankingSection: SortedSection = { departure, arrival, sorted: true };
+              const facilities = canonicalFacilitySections(
+                rankingLine,
+                rankingSection,
+                rankingFacility,
+              );
+              if (facilities === undefined) return [];
+              const rankingQuote = createQuote({
+                version,
+                campaign,
+                line: rankingLine,
+                section: rankingSection,
+                ...facilities,
+                season,
+              });
+              if (rankingQuote.points === undefined) return [];
+              return [{
+                key: `${rankingGroupName}-${rankingLine.at(-1)!.name}-${departure.name}-${arrival.name}`,
+                group: rankingGroupName,
+                departure: departure.name,
+                arrival: arrival.name,
+                ...rankingQuote,
+                value:
+                  rankingFacility === "ordinary" &&
+                  ordinaryRankingBasis === "nonReserved"
+                    ? rate(
+                        rankingQuote.nonReservedFare ?? rankingQuote.paperFare,
+                        rankingQuote.points,
+                      ) ?? 0
+                    : rate(rankingQuote.paperFare, rankingQuote.points) ?? 0,
+              }];
+            }),
+          ),
+        ),
+    );
+    return rows.sort((a, b) => b.value - a.value);
+  }, [campaign, ordinaryRankingBasis, rankingFacility, season, tab, version]);
+  const ranking =
+    rankingLimit === "all"
+      ? rankingRows
+      : rankingRows.slice(0, rankingLimit);
+
+  const onVersionChange = (next: DataVersion) => {
+    setVersion(next);
+    setCampaign("regular");
+    if (next === "2022-03-12") {
+      setRankingFacility("ordinary");
+      setGreenEnabled(false);
+      setGranClassEnabled(false);
+      setRefreshments(false);
+    }
+  };
+
+  const onCampaignChange = (next: Campaign) => {
+    setCampaign(next);
+    if (next === "shinshuPreDc") {
+      setRankingFacility("ordinary");
+      setGreenEnabled(false);
+      setGranClassEnabled(false);
+      setRefreshments(false);
+    }
+  };
+
+  const onFacilityChange = (next: Facility) => {
+    const usesGreen = next !== "ordinary";
+    const usesGranClass =
+      next === "granClassNoRefreshments" ||
+      next === "granClassWithRefreshments";
+    setGreenEnabled(usesGreen);
+    setGranClassEnabled(usesGranClass);
+    setRefreshments(next === "granClassWithRefreshments");
+    if (usesGreen) {
+      setGreenStart(tripStart);
+      setGreenEnd(tripEnd);
+    }
+    if (usesGranClass) {
+      setGranClassStart(tripStart);
+      setGranClassEnd(Math.min(tripEnd, granClassLimit));
+    }
+  };
 
   return (
-    <>
-      <Navbar variant="dark" bg="dark" expand="lg">
-        <Container>
-          <Navbar.Brand>JRE POINT特典チケットのレート計算</Navbar.Brand>
-          <Navbar.Toggle aria-controls="basic-navbar-nav" />
-          <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="me-auto">
-              <Nav.Link
-                as="button"
-                active={tab === "detail"}
-                onClick={() => setTab("detail")}
-              >
-                区間を指定して調べる
-              </Nav.Link>
-              <Nav.Link
-                as="button"
-                active={tab === "ranking"}
-                onClick={() => setTab("ranking")}
-              >
-                ランキング
-              </Nav.Link>
-            </Nav>
-            <Nav>
-              <Nav.Link
-                href="https://github.com/hiroto7/shinkansen"
-                target="_blank"
-                rel="noreferrer"
-              >
-                <i className="bi bi-github" /> GitHub
-              </Nav.Link>
-            </Nav>
-          </Navbar.Collapse>
-        </Container>
-      </Navbar>
-      <Container>
-        <Alert variant="warning" className="mt-3">
-          <h2 className="alert-heading h4">
-            このサイトは更新を終了しています
-          </h2>
-          <p className="mb-0">
-            このサイトの掲載内容・計算結果は、2022年3月12日改正時点の情報をもとにしており、その後の運賃改定、旅客営業規則、JRE
-            POINT特典チケットの交換ポイントなどの変更は反映していません。掲載内容は、現行の運賃・制度とは異なります。正確性を保証しません。最新の情報はご自身でお調べください。
-          </p>
-        </Alert>
-        <div className="mt-4">
-          {tab === "ranking" ? (
-            <Ranking
-              rankedFares={rankedFares}
-              season={season}
-              pointTicketType={pointTicketType}
-              onSeasonChange={setSeason}
-              onPointTicketTypeChange={setPointTicketType}
-            />
-          ) : (
-            <Home
-              season={season}
-              pointTicketType={pointTicketType}
-              onSeasonChange={setSeason}
-              onPointTicketTypeChange={setPointTicketType}
-              rankedFares={rankedFares}
-            />
-          )}
-        </div>
-      </Container>
-    </>
+    <div className="app-shell">
+      <header className="site-header">
+        <h1>新幹線特典レート</h1>
+        <a href="https://github.com/hiroto7/shinkansen">GitHub</a>
+      </header>
+
+      <main>
+        <nav className="tabs" aria-label="表示切替" role="tablist">
+          <button
+            type="button"
+            role="tab"
+            id="detail-tab"
+            aria-controls="detail-panel"
+            aria-selected={tab === "detail"}
+            className={tab === "detail" ? "active" : ""}
+            onClick={() => setTab("detail")}
+          >
+            区間を調べる
+          </button>
+          <button
+            type="button"
+            role="tab"
+            id="ranking-tab"
+            aria-controls="ranking-panel"
+            aria-selected={tab === "ranking"}
+            className={tab === "ranking" ? "active" : ""}
+            onClick={() => setTab("ranking")}
+          >
+            ランキング
+          </button>
+        </nav>
+
+        <section className="control-bar">
+          <label>データ年版
+            <select value={version} onChange={(event) => onVersionChange(event.target.value as DataVersion)}>
+              {versionOptions.map((option) => <option key={option.value} value={option.value}>{option.label}</option>)}
+            </select>
+          </label>
+          <label>交換レート
+            <select value={campaign} onChange={(event) => onCampaignChange(event.target.value as Campaign)}>
+              <option value="regular">通常</option>
+              {version === "2022-03-12" ? (
+                <option value="shinkansenYear">新幹線YEARスペシャル</option>
+              ) : (
+                <>
+                  <option value="limited35Percent">全線35%特別レート</option>
+                  <option value="shinshuPreDc">信州プレDC（対象区間のみ）</option>
+                </>
+              )}
+            </select>
+          </label>
+          <label>シーズン
+            <select value={season} onChange={(event) => setSeason(event.target.value as Season)}>
+              {legacy2022Engine.seasons.map((value) => <option key={value} value={value}>{seasonLabels[value]}</option>)}
+            </select>
+          </label>
+        </section>
+
+        {tab === "detail" ? (
+          <div
+            className="dashboard-grid"
+            role="tabpanel"
+            id="detail-panel"
+            aria-labelledby="detail-tab"
+          >
+            <section className="panel journey-panel">
+              <div className="panel-heading"><h3>乗車区間</h3></div>
+              <div className="field-grid">
+                <label>路線
+                  <select value={groupName} onChange={(event) => resetRoute(event.target.value)}>
+                    {[...legacy2022Engine.lineGroups.keys()].map((name) => <option key={name}>{name}</option>)}
+                  </select>
+                </label>
+                {group.lines.length > 1 && (
+                  <label>方面
+                    <select value={routeIndex} onChange={(event) => resetRoute(groupName, Number(event.target.value))}>
+                      {group.lines.map((route, index) => <option key={route.at(-1)!.name} value={index}>{route.at(-1)!.name}方面</option>)}
+                    </select>
+                  </label>
+                )}
+              </div>
+              <RangeSelect label="全乗車区間" stations={line} start={departureIndex} end={arrivalIndex} onStart={setDepartureIndex} onEnd={setArrivalIndex} />
+
+              {(line === legacy2022Engine.line0 || line === legacy2022Engine.line1) && (
+                <div className="segment-control">
+                  <label className="switch"><input type="checkbox" checked={highSpeedEnabled} onChange={(event) => { setHighSpeedEnabled(event.target.checked); if (event.target.checked) { setHighSpeedStart(tripStart); setHighSpeedEnd(tripEnd); } }} /><span>「はやぶさ」「こまち」を利用する</span></label>
+                  {highSpeedEnabled && highSpeed && <RangeSelect label="はやぶさ・こまち利用区間" stations={tripStations} start={highSpeed.start} end={highSpeed.end} onStart={setHighSpeedStart} onEnd={setHighSpeedEnd} />}
+                </div>
+              )}
+
+              {version === "2026-03-14" && (
+                <div className="facility-stack">
+                  <fieldset className="facility-picker">
+                    <legend>利用する最上位の座席設備</legend>
+                    {([
+                      ["ordinary", "普通車指定席", "全区間で普通車指定席を利用"],
+                      ["green", "グリーン車", "一部または全部で利用"],
+                      ["granClassNoRefreshments", "グランクラス", "飲料・軽食なし"],
+                      ["granClassWithRefreshments", "グランクラス", "飲料・軽食あり"],
+                    ] as const).map(([value, title, description]) => {
+                      const isGranClass = value.startsWith("granClass");
+                      const campaignUnavailable =
+                        campaign === "shinshuPreDc" && value !== "ordinary";
+                      return (
+                        <label className="facility-option" key={value}>
+                          <input
+                            type="radio"
+                            name="facility"
+                            value={value}
+                            checked={selectedFacility === value}
+                            disabled={campaignUnavailable || (isGranClass && !granClassAvailable)}
+                            onChange={() => onFacilityChange(value)}
+                          />
+                          <strong>{title}</strong>
+                          <span>
+                            {campaignUnavailable
+                              ? "信州プレDC対象外"
+                              : isGranClass && !granClassAvailable
+                                ? "この区間では選択不可"
+                                : description}
+                          </span>
+                        </label>
+                      );
+                    })}
+                  </fieldset>
+                  {greenEnabled && (
+                    <div className="segment-control">
+                      <RangeSelect
+                        label={granClassEnabled ? "グリーン車・グランクラス利用区間" : "グリーン車利用区間"}
+                        stations={tripStations}
+                        start={green?.start ?? tripStart}
+                        end={green?.end ?? tripEnd}
+                        onStart={setGreenStart}
+                        onEnd={setGreenEnd}
+                      />
+                      {granClassEnabled && granClassAvailable && (
+                        <RangeSelect
+                          label="グランクラス利用区間"
+                          stations={granClassStations}
+                          start={granClass?.start ?? granClassOuterStart}
+                          end={granClass?.end ?? granClassOuterEnd}
+                          onStart={setGranClassStart}
+                          onEnd={setGranClassEnd}
+                        />
+                      )}
+                    </div>
+                  )}
+                </div>
+              )}
+              <p className="constraint-note">設備ごとの利用区間は、それぞれ1つの連続区間として指定してください。</p>
+            </section>
+
+            <section className="panel result-panel">
+              <div className="panel-heading"><h3>計算結果</h3></div>
+              {quote.points === undefined ? (
+                <div className="unavailable">
+                  <strong>
+                    {exclusionMessages[quote.exclusionReason ?? "invalidJourney"].title}
+                  </strong>
+                  <p>{exclusionMessages[quote.exclusionReason ?? "invalidJourney"].detail}</p>
+                </div>
+              ) : (
+                <>
+                  <div className="result-route">
+                    <strong>{sorted.departure.name} → {sorted.arrival.name}</strong>
+                    <span>{quote.distanceKm.toFixed(1)} km</span>
+                  </div>
+                  <section className="result-section" aria-labelledby="points-heading">
+                    <h4 id="points-heading">必要ポイント</h4>
+                    <div className="points-total"><span>{facilityLabels[quote.facility]}</span><strong>{integer.format(quote.points)}<small> pt</small></strong></div>
+                  </section>
+                  <section className="result-section" aria-labelledby="comparison-heading">
+                    <h4 id="comparison-heading">所定額との比較</h4>
+                    <div className={`rate-grid${quote.facility === "ordinary" && quote.nonReservedFare !== undefined ? "" : " single"}`}>
+                      {quote.facility === "ordinary" && quote.nonReservedFare !== undefined && (
+                        <RateCard label="自由席・立席" fare={quote.nonReservedFare} points={quote.points} />
+                      )}
+                      <RateCard
+                        label={quote.facility === "ordinary" ? "普通車指定席" : "同じ設備・行程"}
+                        fare={quote.paperFare}
+                        points={quote.points}
+                      />
+                    </div>
+                  </section>
+                  <section className="result-section" aria-labelledby="breakdown-heading">
+                    <h4 id="breakdown-heading">{facilityLabels[quote.facility]}の所定額内訳</h4>
+                    <dl className="breakdown">
+                      <div><dt>普通運賃</dt><dd>{quote.basicFare === undefined ? "—" : yen.format(quote.basicFare)}</dd></div>
+                      <div><dt>特急料金</dt><dd>{quote.expressFare === undefined ? "—" : yen.format(quote.expressFare)}</dd></div>
+                      {(quote.specialVehicleFare ?? 0) > 0 && (
+                        <div><dt>特別車両料金</dt><dd>{yen.format(quote.specialVehicleFare!)}</dd></div>
+                      )}
+                      <div className="total"><dt>合計</dt><dd>{quote.paperFare === undefined ? "—" : yen.format(quote.paperFare)}</dd></div>
+                    </dl>
+                  </section>
+                </>
+              )}
+            </section>
+          </div>
+        ) : (
+          <section
+            className="panel ranking-panel"
+            role="tabpanel"
+            id="ranking-panel"
+            aria-labelledby="ranking-tab"
+          >
+            <div className="panel-heading"><h3>{facilityLabels[rankingFacility]}レート</h3></div>
+            <div className="ranking-tools">
+              <label>設備
+                <select value={rankingFacility} onChange={(event) => setRankingFacility(event.target.value as RankingFacility)}>
+                  <option value="ordinary">普通車指定席</option>
+                  {version === "2026-03-14" && campaign !== "shinshuPreDc" && <>
+                    <option value="green">グリーン車</option>
+                    <option value="granClassNoRefreshments">グランクラス（飲料・軽食なし）</option>
+                    <option value="granClassWithRefreshments">グランクラス（飲料・軽食あり）</option>
+                  </>}
+                </select>
+              </label>
+              {rankingFacility === "ordinary" && (
+                <label>並び順
+                  <select value={ordinaryRankingBasis} onChange={(event) => setOrdinaryRankingBasis(event.target.value as OrdinaryRankingBasis)}>
+                    <option value="nonReserved">自由席・立席との比較</option>
+                    <option value="reserved">普通車指定席との比較</option>
+                  </select>
+                </label>
+              )}
+              <label>表示件数
+                <select value={rankingLimit} onChange={(event) => setRankingLimit(event.target.value === "all" ? "all" : Number(event.target.value) as 50 | 100)}>
+                  <option value={50}>50件</option>
+                  <option value={100}>100件</option>
+                  <option value="all">全件</option>
+                </select>
+              </label>
+            </div>
+            <p className="ranking-note">全{integer.format(rankingRows.length)}件中、{integer.format(ranking.length)}件を表示。{rankingFacility === "ordinary" ? `自由席・立席と普通車指定席の所定額を比較できます。${ordinaryRankingBasis === "nonReserved" ? "自由席・立席を設定しない区間は、指定席レートで並べます。" : ""}` : "指定した設備と同じ行程の所定額との比較です。"}</p>
+            <div className="ranking-table-wrap"><table className={`ranking-table${rankingFacility === "ordinary" ? " ordinary" : ""}`}><thead><tr><th>#</th><th>区間</th><th>距離</th><th>ポイント</th>{rankingFacility === "ordinary" ? <><th>自由席・立席</th><th>円/pt</th><th>普通車指定席</th><th>円/pt</th></> : <><th>所定額</th><th>円/pt</th></>}</tr></thead><tbody>
+              {ranking.map((row, index) => <tr key={row.key}><td>{index + 1}</td><td><small>{row.group}</small><strong>{row.departure} → {row.arrival}</strong></td><td>{row.distanceKm.toFixed(1)} km</td><td>{row.points === undefined ? "—" : integer.format(row.points)}</td>{rankingFacility === "ordinary" ? <><td>{row.nonReservedFare === undefined ? "—" : yen.format(row.nonReservedFare)}</td><td>{row.nonReservedFare === undefined ? "—" : <strong className={ordinaryRankingBasis === "nonReserved" ? "selected-rate" : ""}>{(rate(row.nonReservedFare, row.points) ?? 0).toFixed(2)}</strong>}</td><td>{row.paperFare === undefined ? "—" : yen.format(row.paperFare)}</td><td><strong className={ordinaryRankingBasis === "reserved" ? "selected-rate" : ""}>{(rate(row.paperFare, row.points) ?? 0).toFixed(2)}</strong></td></> : <><td>{row.paperFare === undefined ? "—" : yen.format(row.paperFare)}</td><td><strong className="selected-rate">{row.value.toFixed(2)}</strong></td></>}</tr>)}
+            </tbody></table></div>
+          </section>
+        )}
+
+        <aside className="notice">
+          <strong>計算の前提</strong>
+          <p>運賃・料金はJRの規則と公式表、交換ポイントはえきねっとの公式表から計算します。時刻表・列車編成・残席・実際の発売可否は判定しません。表示額は購入を保証するものではありません。最新情報はご自身でお調べください。</p>
+        </aside>
+      </main>
+    </div>
   );
-};
-
-export type {
-  ExpressTicket,
-  Line,
-  LineGroup,
-  PointTicketType,
-  Season,
-  Section,
-  SortedSection,
-  Station,
-  TotalFare,
-};
-
-export const legacy2022Engine: {
-  readonly lineGroups: ReadonlyMap<string, LineGroup>;
-  readonly junctions: ReadonlyMap<Line, Station>;
-  readonly line0: Line;
-  readonly line1: Line;
-  readonly line2: Line;
-  readonly line3: Line;
-  readonly line4: Line;
-  readonly line5: Line;
-  readonly average: Season;
-  readonly busy: Season;
-  readonly busiest: Season;
-  readonly off: Season;
-  readonly seasons: readonly Season[];
-  readonly pointTicketTypes: readonly PointTicketType[];
-  readonly getFares: typeof getFares;
-  readonly sortSection: typeof sortSection;
-} = {
-  lineGroups,
-  junctions,
-  line0,
-  line1,
-  line2,
-  line3,
-  line4,
-  line5,
-  average,
-  busy,
-  busiest,
-  off,
-  seasons,
-  pointTicketTypes,
-  getFares,
-  sortSection,
 };
 
 export default App;
