+++ src/index.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./modern/App";
+import App from "./App";
 import "./index.css";
 
 const rootElement = document.getElementById("root");
